Network.AWS.SWF.CountClosedWorkflowExecutions

Description

Returns the number of closed workflow executions within the given domain that meet the specified filtering criteria. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. tagFilter.tag: String constraint. The key is swf:tagFilter.tag. typeFilter.name: String constraint. The key is swf:typeFilter.name. typeFilter.version: String constraint. The key is swf:typeFilter.version. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_CountClosedWorkflowExecutions.html

Request

data CountClosedWorkflowExecutions

Request constructor

countClosedWorkflowExecutions

Arguments

:: Text
ccweDomain

-> CountClosedWorkflowExecutions

CountClosedWorkflowExecutions constructor.

The fields accessible through corresponding lenses are:

ccweCloseStatusFilter :: Maybe CloseStatusFilter

ccweCloseTimeFilter :: Maybe ExecutionTimeFilter

ccweDomain :: Text

ccweExecutionFilter :: Maybe WorkflowExecutionFilter

ccweStartTimeFilter :: Maybe ExecutionTimeFilter

ccweTagFilter :: Maybe TagFilter

ccweTypeFilter :: Maybe WorkflowTypeFilter

Request lenses

ccweCloseStatusFilter :: Lens' CountClosedWorkflowExecutions (Maybe CloseStatusFilter)

If specified, only workflow executions that match this close status are counted. This filter has an affect only if executionStatus is specified as CLOSED.

ccweCloseTimeFilter :: Lens' CountClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, only workflow executions that meet the close time criteria of the filter are counted.

ccweDomain :: Lens' CountClosedWorkflowExecutions Text

The name of the domain containing the workflow executions to count.

ccweExecutionFilter :: Lens' CountClosedWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the WorkflowId in the filter are counted.

ccweStartTimeFilter :: Lens' CountClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, only workflow executions that meet the start time criteria of the filter are counted.

ccweTagFilter :: Lens' CountClosedWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have a tag that matches the filter are counted.

ccweTypeFilter :: Lens' CountClosedWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, indicates the type of the workflow executions to be counted.

Response

data CountClosedWorkflowExecutionsResponse

Response constructor

countClosedWorkflowExecutionsResponse

Arguments

:: Natural
ccwerCount

-> CountClosedWorkflowExecutionsResponse

CountClosedWorkflowExecutionsResponse constructor.

The fields accessible through corresponding lenses are:

ccwerCount :: Natural

ccwerTruncated :: Maybe Bool

Response lenses

ccwerCount :: Lens' CountClosedWorkflowExecutionsResponse Natural

The number of workflow executions.

ccwerTruncated :: Lens' CountClosedWorkflowExecutionsResponse (Maybe Bool)

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.