Network.AWS.SWF.CountOpenWorkflowExecutions

Description

Returns the number of open workflow executions within the given domain that meet the specified filtering criteria.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. 'tagFilter.tag': String constraint. The key is 'swf:tagFilter.tag'. 'typeFilter.name': String constraint. The key is 'swf:typeFilter.name'. 'typeFilter.version': String constraint. The key is 'swf:typeFilter.version'. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to AmazonSWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_CountOpenWorkflowExecutions.html

Request

data CountOpenWorkflowExecutions

Request constructor

countOpenWorkflowExecutions

Arguments

:: Text
coweDomain

-> ExecutionTimeFilter
coweStartTimeFilter

-> CountOpenWorkflowExecutions

CountOpenWorkflowExecutions constructor.

The fields accessible through corresponding lenses are:

coweDomain :: Text

coweExecutionFilter :: Maybe WorkflowExecutionFilter

coweStartTimeFilter :: ExecutionTimeFilter

coweTagFilter :: Maybe TagFilter

coweTypeFilter :: Maybe WorkflowTypeFilter

Request lenses

coweDomain :: Lens' CountOpenWorkflowExecutions Text

The name of the domain containing the workflow executions to count.

coweExecutionFilter :: Lens' CountOpenWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the WorkflowId in the filter are counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

coweStartTimeFilter :: Lens' CountOpenWorkflowExecutions ExecutionTimeFilter

Specifies the start time criteria that workflow executions must meet in order to be counted.

coweTagFilter :: Lens' CountOpenWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have a tag that matches the filter are counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

coweTypeFilter :: Lens' CountOpenWorkflowExecutions (Maybe WorkflowTypeFilter)

Specifies the type of the workflow executions to be counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

Response

data CountOpenWorkflowExecutionsResponse

Response constructor

countOpenWorkflowExecutionsResponse

Arguments

:: Natural
cowerCount

-> CountOpenWorkflowExecutionsResponse

CountOpenWorkflowExecutionsResponse constructor.

The fields accessible through corresponding lenses are:

cowerCount :: Natural

cowerTruncated :: Maybe Bool

Response lenses

cowerCount :: Lens' CountOpenWorkflowExecutionsResponse Natural

The number of workflow executions.

cowerTruncated :: Lens' CountOpenWorkflowExecutionsResponse (Maybe Bool)

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.