Network.AWS.SWF.DeprecateDomain

Description

Deprecates the specified domain. After a domain has been deprecated it cannot be used to create new workflow executions or register new types. However, you can still use visibility actions on this domain. Deprecating a domain also deprecates all activity and workflow types registered in the domain. Executions that were started before the domain was deprecated will continue to run.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_DeprecateDomain.html

Request

data DeprecateDomain

Request constructor

deprecateDomain

Arguments

:: Text
dd1Name

-> DeprecateDomain

DeprecateDomain constructor.

The fields accessible through corresponding lenses are:

dd1Name :: Text

Request lenses

dd1Name :: Lens' DeprecateDomain Text

The name of the domain to deprecate.

Response

data DeprecateDomainResponse

Response constructor

deprecateDomainResponse :: DeprecateDomainResponse

DeprecateDomainResponse constructor.