Network.AWS.SWF.DeprecateWorkflowType

Description

Deprecates the specified workflow type. After a workflow type has been deprecated, you cannot create new executions of that type. Executions that were started before the type was deprecated will continue to run. A deprecated workflow type may still be used when calling visibility actions.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. 'workflowType.name': String constraint. The key is 'swf:workflowType.name'. 'workflowType.version': String constraint. The key is 'swf:workflowType.version' . If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_DeprecateWorkflowType.html

Request

data DeprecateWorkflowType

Request constructor

deprecateWorkflowType

Arguments

:: Text
dwt1Domain

-> WorkflowType
dwt1WorkflowType

-> DeprecateWorkflowType

DeprecateWorkflowType constructor.

The fields accessible through corresponding lenses are:

dwt1Domain :: Text

dwt1WorkflowType :: WorkflowType

Request lenses

dwt1Domain :: Lens' DeprecateWorkflowType Text

The name of the domain in which the workflow type is registered.

dwt1WorkflowType :: Lens' DeprecateWorkflowType WorkflowType

The workflow type to deprecate.

Response

data DeprecateWorkflowTypeResponse

Response constructor

deprecateWorkflowTypeResponse :: DeprecateWorkflowTypeResponse

DeprecateWorkflowTypeResponse constructor.