Network.AWS.SWF.DescribeDomain

Description

Returns information about the specified domain, including description and status.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

describeDomain

Arguments

:: Text
ddName

-> DescribeDomain

Creates a value of DescribeDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddName - The name of the domain to describe.

data DescribeDomain

See: describeDomain smart constructor.

Request Lenses

ddName :: Lens' DescribeDomain Text

The name of the domain to describe.

Destructuring the Response

describeDomainResponse

Arguments

:: Int
ddrsResponseStatus

-> DomainInfo
ddrsDomainInfo

-> DomainConfiguration
ddrsConfiguration

-> DescribeDomainResponse

Creates a value of DescribeDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddrsResponseStatus - -- | The response status code.

ddrsDomainInfo - The basic information about a domain, such as its name, status, and description.

ddrsConfiguration - The domain configuration. Currently, this includes only the domain's retention period.

data DescribeDomainResponse

Contains details of a domain.

See: describeDomainResponse smart constructor.

Response Lenses

ddrsResponseStatus :: Lens' DescribeDomainResponse Int

- | The response status code.

ddrsDomainInfo :: Lens' DescribeDomainResponse DomainInfo

The basic information about a domain, such as its name, status, and description.

ddrsConfiguration :: Lens' DescribeDomainResponse DomainConfiguration

The domain configuration. Currently, this includes only the domain's retention period.