Network.AWS.SWF.DescribeWorkflowExecution

Description

Returns information about the specified workflow execution including its type and some statistics.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_DescribeWorkflowExecution.html

Request

data DescribeWorkflowExecution

Request constructor

describeWorkflowExecution

Arguments

:: Text
dweDomain

-> WorkflowExecution
dweExecution

-> DescribeWorkflowExecution

DescribeWorkflowExecution constructor.

The fields accessible through corresponding lenses are:

dweDomain :: Text

dweExecution :: WorkflowExecution

Request lenses

dweDomain :: Lens' DescribeWorkflowExecution Text

The name of the domain containing the workflow execution.

dweExecution :: Lens' DescribeWorkflowExecution WorkflowExecution

The workflow execution to describe.

Response

data DescribeWorkflowExecutionResponse

Response constructor

describeWorkflowExecutionResponse

Arguments

:: WorkflowExecutionInfo
dwerExecutionInfo

-> WorkflowExecutionConfiguration
dwerExecutionConfiguration

-> WorkflowExecutionOpenCounts
dwerOpenCounts

-> DescribeWorkflowExecutionResponse

DescribeWorkflowExecutionResponse constructor.

The fields accessible through corresponding lenses are:

dwerExecutionConfiguration :: WorkflowExecutionConfiguration

dwerExecutionInfo :: WorkflowExecutionInfo

dwerLatestActivityTaskTimestamp :: Maybe UTCTime

dwerLatestExecutionContext :: Maybe Text

dwerOpenCounts :: WorkflowExecutionOpenCounts

Response lenses

dwerExecutionConfiguration :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionConfiguration

The configuration settings for this workflow execution including timeout values, tasklist etc.

dwerExecutionInfo :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionInfo

Information about the workflow execution.

dwerLatestActivityTaskTimestamp :: Lens' DescribeWorkflowExecutionResponse (Maybe UTCTime)

The time when the last activity task was scheduled for this workflow execution. You can use this information to determine if the workflow has not made progress for an unusually long period of time and might require a corrective action.

dwerLatestExecutionContext :: Lens' DescribeWorkflowExecutionResponse (Maybe Text)

The latest executionContext provided by the decider for this workflow execution. A decider can provide an executionContext, which is a free form string, when closing a decision task using RespondDecisionTaskCompleted.

dwerOpenCounts :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionOpenCounts

The number of tasks for this workflow execution. This includes open and closed tasks of all types.