Network.AWS.SWF.DescribeWorkflowType

Description

Returns information about the specified workflow type. This includes configuration settings specified when the type was registered and other information such as creation date, current status, etc. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. workflowType.name: String constraint. The key is swf:workflowType.name. workflowType.version: String constraint. The key is swf:workflowType.version. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows. DescribeWorkflowType Example POST HTTP1.1 Host: swf.us-east-1.amazonaws.com User-Agent: Mozilla/5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.2.25) Gecko/20111212 Firefox3.6.25 ( .NET CLR 3.5.30729; .NET4.0E) Accept: applicationjson, textjavascript, ** Accept-Language: en-us,en;q=0.5 Accept-Encoding: gzip,deflate Accept-Charset: ISO-8859-1,utf-8;q=0.7,*;q=0.7 Keep-Alive: 115 Connection: keep-alive Content-Type: application/x-amz-json-1.0 X-Requested-With: XMLHttpRequest X-Amz-Date: Sun, 15 Jan 2012 22:40:40 GMT X-Amz-Target: SimpleWorkflowService.DescribeWorkflowType Content-Encoding: amz-1.0 X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE,Algorithm=HmacSHA256,SignedHeaders=Host;X-Amz-Date;X-Amz-Target;Content-Encoding,Signature=iGt8t83OmrURqu0pKYbcW6mNdjXbFomevCBPUPQEbaM= Referer: http:/swf.us-east-1.amazonaws.comexplorer/index.html Content-Length: 102 Pragma: no-cache Cache-Control: no-cache {"domain": "867530901", "workflowType": {"name": "customerOrderWorkflow", "version": "1.0"} } HTTP1.1 200 OK Content-Length: 348 Content-Type: applicationjson x-amzn-RequestId: f35a8e7f-3fc9-11e1-a23a-99d60383ae71 {"configuration": {"defaultChildPolicy": TERMINATE, "defaultExecutionStartToCloseTimeout": "3600", "defaultTaskList": {"name": "mainTaskList"}, "defaultTaskStartToCloseTimeout": "600"}, "typeInfo": {"creationDate": 1326481174.027, "description": "Handle customer orders", "status": REGISTERED, "workflowType": {"name": "customerOrderWorkflow", "version": "1.0"} } }.

Request

data DescribeWorkflowType

Request constructor

describeWorkflowType

Arguments

:: Text
dwt1Domain

-> WorkflowType
dwt1WorkflowType

-> DescribeWorkflowType

Smart constructor for the minimum required parameters to construct a valid DescribeWorkflowType request.

The fields accessible through corresponding lenses are:

Domain :: Text

WorkflowType :: WorkflowType

Request lenses

dwt1Domain :: Lens' DescribeWorkflowType Text

The name of the domain in which this workflow type is registered.

dwt1WorkflowType :: Lens' DescribeWorkflowType WorkflowType

The workflow type to describe.

Response

data DescribeWorkflowTypeResponse

Contains details about a workflow type.

Response constructor

describeWorkflowTypeResponse

Arguments

:: WorkflowTypeInfo
dwtrTypeInfo

-> WorkflowTypeConfiguration
dwtrConfiguration

-> DescribeWorkflowTypeResponse

Smart constructor for the minimum required parameters to construct a valid DescribeWorkflowTypeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TypeInfo :: WorkflowTypeInfo

Configuration :: WorkflowTypeConfiguration

Response lenses

dwtrTypeInfo :: Lens' DescribeWorkflowTypeResponse WorkflowTypeInfo

General information about the workflow type. The status of the workflow type (returned in the WorkflowTypeInfo structure) can be one of the following. REGISTERED: The type is registered and available. Workers supporting this type should be running. DEPRECATED: The type was deprecated using DeprecateWorkflowType, but is still in use. You should keep workers supporting this type running. You cannot create new workflow executions of this type.

dwtrConfiguration :: Lens' DescribeWorkflowTypeResponse WorkflowTypeConfiguration

Configuration settings of the workflow type registered through RegisterWorkflowType.