Network.AWS.SWF.GetWorkflowExecutionHistory

Description

Returns the history of the specified workflow execution. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see <http://docs.aws.amazon.com/amazonswf/latest/developerguide/swf-dev-iam.html Using IAM to Manage Access to Amazon SWF Workflows>.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_GetWorkflowExecutionHistory.html

Request

data GetWorkflowExecutionHistory

Request constructor

getWorkflowExecutionHistory

Arguments

:: Text
gwehDomain

-> WorkflowExecution
gwehExecution

-> GetWorkflowExecutionHistory

GetWorkflowExecutionHistory constructor.

The fields accessible through corresponding lenses are:

gwehDomain :: Text

gwehExecution :: WorkflowExecution

gwehMaximumPageSize :: Maybe Natural

gwehNextPageToken :: Maybe Text

gwehReverseOrder :: Maybe Bool

Request lenses

gwehDomain :: Lens' GetWorkflowExecutionHistory Text

The name of the domain containing the workflow execution.

gwehExecution :: Lens' GetWorkflowExecutionHistory WorkflowExecution

Specifies the workflow execution for which to return the history.

gwehMaximumPageSize :: Lens' GetWorkflowExecutionHistory (Maybe Natural)

Specifies the maximum number of history events returned in one page. The next page in the result is identified by the NextPageToken returned. By default 100 history events are returned in a page but the caller can override this value to a page size smaller than the default. You cannot specify a page size larger than 100. Note that the number of events may be less than the maxiumum page size, in which case, the returned page will have fewer results than the maximumPageSize specified.

gwehNextPageToken :: Lens' GetWorkflowExecutionHistory (Maybe Text)

If a NextPageToken is returned, the result has more than one pages. To get the next page, repeat the call and specify the nextPageToken with all other arguments unchanged.

gwehReverseOrder :: Lens' GetWorkflowExecutionHistory (Maybe Bool)

When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimeStamp of the events.

Response

data GetWorkflowExecutionHistoryResponse

Response constructor

getWorkflowExecutionHistoryResponse :: GetWorkflowExecutionHistoryResponse

GetWorkflowExecutionHistoryResponse constructor.

The fields accessible through corresponding lenses are:

gwehrEvents :: [HistoryEvent]

gwehrNextPageToken :: Maybe Text

Response lenses

gwehrEvents :: Lens' GetWorkflowExecutionHistoryResponse [HistoryEvent]

The list of history events.

gwehrNextPageToken :: Lens' GetWorkflowExecutionHistoryResponse (Maybe Text)

The token for the next page. If set, the history consists of more than one page and the next page can be retrieved by repeating the request with this token and all other arguments unchanged.