Network.AWS.SWF.ListClosedWorkflowExecutions

Description

Returns a list of closed workflow executions in the specified domain that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. 'tagFilter.tag': String constraint. The key is 'swf:tagFilter.tag'. 'typeFilter.name': String constraint. The key is 'swf:typeFilter.name'. 'typeFilter.version': String constraint. The key is 'swf:typeFilter.version'. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access toAmazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_ListClosedWorkflowExecutions.html

Request

data ListClosedWorkflowExecutions

Request constructor

listClosedWorkflowExecutions

Arguments

:: Text
lcweDomain

-> ListClosedWorkflowExecutions

ListClosedWorkflowExecutions constructor.

The fields accessible through corresponding lenses are:

lcweCloseStatusFilter :: Maybe CloseStatusFilter

lcweCloseTimeFilter :: Maybe ExecutionTimeFilter

lcweDomain :: Text

lcweExecutionFilter :: Maybe WorkflowExecutionFilter

lcweMaximumPageSize :: Maybe Natural

lcweNextPageToken :: Maybe Text

lcweReverseOrder :: Maybe Bool

lcweStartTimeFilter :: Maybe ExecutionTimeFilter

lcweTagFilter :: Maybe TagFilter

lcweTypeFilter :: Maybe WorkflowTypeFilter

Request lenses

lcweCloseStatusFilter :: Lens' ListClosedWorkflowExecutions (Maybe CloseStatusFilter)

If specified, only workflow executions that match this close status are listed. For example, if TERMINATED is specified, then only TERMINATED workflow executions are listed.

lcweCloseTimeFilter :: Lens' ListClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, the workflow executions are included in the returned results based on whether their close times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their close times.

lcweDomain :: Lens' ListClosedWorkflowExecutions Text

The name of the domain that contains the workflow executions to list.

lcweExecutionFilter :: Lens' ListClosedWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the workflow id specified in the filter are returned.

lcweMaximumPageSize :: Lens' ListClosedWorkflowExecutions (Maybe Natural)

The maximum number of results returned in each page. The default is 100, but the caller can override this value to a page size smaller than the default. You cannot specify a page size greater than 100. Note that the number of executions may be less than the maxiumum page size, in which case, the returned page will have fewer results than the maximumPageSize specified.

lcweNextPageToken :: Lens' ListClosedWorkflowExecutions (Maybe Text)

If on a previous call to this method a NextPageToken was returned, the results are being paginated. To get the next page of results, repeat the call with the returned token and all other arguments unchanged.

lcweReverseOrder :: Lens' ListClosedWorkflowExecutions (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in descending order of the start or the close time of the executions.

lcweStartTimeFilter :: Lens' ListClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, the workflow executions are included in the returned results based on whether their start times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their start times.

lcweTagFilter :: Lens' ListClosedWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have the matching tag are listed.

lcweTypeFilter :: Lens' ListClosedWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, only executions of the type specified in the filter are returned.

Response

data ListClosedWorkflowExecutionsResponse

Response constructor

listClosedWorkflowExecutionsResponse :: ListClosedWorkflowExecutionsResponse

ListClosedWorkflowExecutionsResponse constructor.

The fields accessible through corresponding lenses are:

lcwerExecutionInfos :: [WorkflowExecutionInfo]

lcwerNextPageToken :: Maybe Text

Response lenses

lcwerExecutionInfos :: Lens' ListClosedWorkflowExecutionsResponse [WorkflowExecutionInfo]

The list of workflow information structures.

lcwerNextPageToken :: Lens' ListClosedWorkflowExecutionsResponse (Maybe Text)

The token of the next page in the result. If set, the results have more than one page. The next page can be retrieved by repeating the request with this token and all other arguments unchanged.