Network.AWS.SWF.ListDomains

Description

Returns the list of domains registered in the account. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. The element must be set to 'arn:aws:swf::AccountID:domain/*', where AccountID is the account ID, with no dashes. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_ListDomains.html

Request

data ListDomains

Request constructor

listDomains

Arguments

:: RegistrationStatus
ldRegistrationStatus

-> ListDomains

ListDomains constructor.

The fields accessible through corresponding lenses are:

ldMaximumPageSize :: Maybe Natural

ldNextPageToken :: Maybe Text

ldRegistrationStatus :: RegistrationStatus

ldReverseOrder :: Maybe Bool

Request lenses

ldMaximumPageSize :: Lens' ListDomains (Maybe Natural)

The maximum number of results that will be returned per call. nextPageToken can be used to obtain futher pages of results. The default is 100, which is the maximum allowed page size. You can, however, specify a page size smaller than 100.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

ldNextPageToken :: Lens' ListDomains (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

ldRegistrationStatus :: Lens' ListDomains RegistrationStatus

Specifies the registration status of the domains to list.

ldReverseOrder :: Lens' ListDomains (Maybe Bool)

When set to true, returns the results in reverse order. By default, the results are returned in ascending alphabetical order by name of the domains.

Response

data ListDomainsResponse

Response constructor

listDomainsResponse :: ListDomainsResponse

ListDomainsResponse constructor.

The fields accessible through corresponding lenses are:

ldrDomainInfos :: [DomainInfo]

ldrNextPageToken :: Maybe Text

Response lenses

ldrDomainInfos :: Lens' ListDomainsResponse [DomainInfo]

A list of DomainInfo structures.

ldrNextPageToken :: Lens' ListDomainsResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.