Network.AWS.SWF.ListWorkflowTypes

Description

Returns information about workflow types in the specified domain. The results may be split into multiple pages that can be retrieved by making the call repeatedly.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_ListWorkflowTypes.html

Request

data ListWorkflowTypes

Request constructor

listWorkflowTypes

Arguments

:: Text
lwtDomain

-> RegistrationStatus
lwtRegistrationStatus

-> ListWorkflowTypes

ListWorkflowTypes constructor.

The fields accessible through corresponding lenses are:

lwtDomain :: Text

lwtMaximumPageSize :: Maybe Natural

lwtName :: Maybe Text

lwtNextPageToken :: Maybe Text

lwtRegistrationStatus :: RegistrationStatus

lwtReverseOrder :: Maybe Bool

Request lenses

lwtDomain :: Lens' ListWorkflowTypes Text

The name of the domain in which the workflow types have been registered.

lwtMaximumPageSize :: Lens' ListWorkflowTypes (Maybe Natural)

The maximum number of results returned in each page. The default is 100, but the caller can override this value to a page size smaller than the default. You cannot specify a page size greater than 100. Note that the number of types may be less than the maxiumum page size, in which case, the returned page will have fewer results than the maximumPageSize specified.

lwtName :: Lens' ListWorkflowTypes (Maybe Text)

If specified, lists the workflow type with this name.

lwtNextPageToken :: Lens' ListWorkflowTypes (Maybe Text)

If on a previous call to this method a NextPageToken was returned, the results are being paginated. To get the next page of results, repeat the call with the returned token and all other arguments unchanged.

lwtRegistrationStatus :: Lens' ListWorkflowTypes RegistrationStatus

Specifies the registration status of the workflow types to list.

lwtReverseOrder :: Lens' ListWorkflowTypes (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in ascending alphabetical order of the name of the workflow types.

Response

data ListWorkflowTypesResponse

Response constructor

listWorkflowTypesResponse :: ListWorkflowTypesResponse

ListWorkflowTypesResponse constructor.

The fields accessible through corresponding lenses are:

lwtrNextPageToken :: Maybe Text

lwtrTypeInfos :: [WorkflowTypeInfo]

Response lenses

lwtrNextPageToken :: Lens' ListWorkflowTypesResponse (Maybe Text)

The token for the next page of type information. If set then the list consists of more than one page. You can retrieve the next page by repeating the request (that returned the structure) with the this token and all other arguments unchanged.

lwtrTypeInfos :: Lens' ListWorkflowTypesResponse [WorkflowTypeInfo]

The list of workflow type information.