Network.AWS.SWF.RecordActivityTaskHeartbeat

Description

Used by activity workers to report to the service that the ActivityTask represented by the specified taskToken is still making progress. The worker can also (optionally) specify details of the progress, for example percent complete, using the details parameter. This action can also be used by the worker as a mechanism to check if cancellation is being requested for the activity task. If a cancellation is being attempted for the specified task, then the boolean cancelRequested flag returned by the service is set to true. This action resets the taskHeartbeatTimeout clock. The taskHeartbeatTimeout is specified in RegisterActivityType. This action does not in itself create an event in the workflow execution history. However, if the task times out, the workflow execution history will contain a ActivityTaskTimedOut event that contains the information from the last heartbeat generated by the activity worker. If the cancelRequested flag returns true, a cancellation is being attempted. If the worker can cancel the activity, it should respond with RespondActivityTaskCanceled. Otherwise, it should ignore the cancellation request. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_RecordActivityTaskHeartbeat.html

Request

data RecordActivityTaskHeartbeat

Request constructor

recordActivityTaskHeartbeat

Arguments

:: Text
rathTaskToken

-> RecordActivityTaskHeartbeat

RecordActivityTaskHeartbeat constructor.

The fields accessible through corresponding lenses are:

rathDetails :: Maybe Text

rathTaskToken :: Text

Request lenses

rathDetails :: Lens' RecordActivityTaskHeartbeat (Maybe Text)

If specified, contains details about the progress of the task.

rathTaskToken :: Lens' RecordActivityTaskHeartbeat Text

The taskToken of the ActivityTask. The taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

Response

data RecordActivityTaskHeartbeatResponse

Response constructor

recordActivityTaskHeartbeatResponse

Arguments

:: Bool
rathrCancelRequested

-> RecordActivityTaskHeartbeatResponse

RecordActivityTaskHeartbeatResponse constructor.

The fields accessible through corresponding lenses are:

rathrCancelRequested :: Bool

Response lenses

rathrCancelRequested :: Lens' RecordActivityTaskHeartbeatResponse Bool

Set to true if cancellation of the task is requested.