Network.AWS.SWF.RegisterDomain

Description

Registers a new domain. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: You cannot use an IAM policy to control domain access for this action. The name of the domain being registered is available as the resource of this action. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows. RegisterDomain Example POST / HTTP1.1 Host: swf.us-east-1.amazonaws.com User-Agent: Mozilla5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.2.25) Gecko/20111212 Firefox3.6.25 ( .NET CLR 3.5.30729; .NET4.0E) Accept: applicationjson, textjavascript, ** Accept-Language: en-us,en;q=0.5 Accept-Encoding: gzip,deflate Accept-Charset: ISO-8859-1,utf-8;q=0.7,*;q=0.7 Keep-Alive: 115 Connection: keep-alive Content-Type: application/x-amz-json-1.0 X-Requested-With: XMLHttpRequest X-Amz-Date: Fri, 13 Jan 2012 18:42:12 GMT X-Amz-Target: SimpleWorkflowService.RegisterDomain Content-Encoding: amz-1.0 X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE,Algorithm=HmacSHA256,SignedHeaders=Host;X-Amz-Date;X-Amz-Target;Content-Encoding,Signature=tzjkF55lxAxPhzpBRGFYQRQRq6CqrM254dTDEEncI= Referer: http:/swf.us-east-1.amazonaws.comexplorer/index.html Content-Length: 91 Pragma: no-cache Cache-Control: no-cache {"name": "867530902", "description": "music", "workflowExecutionRetentionPeriodInDays": "60"} HTTP/1.1 200 OK Content-Length: 0 Content-Type: application/json x-amzn-RequestId: 4ec4ac3f-3e16-11e1-9b11-7182192d0b57.

Request

data RegisterDomain

Request constructor

mkRegisterDomain

Arguments

:: Text
rdName

-> Text
rdWorkflowExecutionRetentionPeriodInDays

-> RegisterDomain

Smart constructor for the minimum required parameters to construct a valid RegisterDomain request.

The fields accessible through corresponding lenses are:

Name :: Text

Description :: Maybe Text

WorkflowExecutionRetentionPeriodInDays :: Text

Request lenses

rdName :: Lens' RegisterDomain Text

Name of the domain to register. The name must be unique. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f - u009f). Also, it must not contain the literal string &quot;arn&quot;.

rdDescription :: Lens' RegisterDomain (Maybe Text)

Textual description of the domain.

rdWorkflowExecutionRetentionPeriodInDays :: Lens' RegisterDomain Text

A duration (in days) for which the record (including the history) of workflow executions in this domain should be kept by the service. After the retention period, the workflow execution will not be available in the results of visibility calls. If you pass the value NONE then there is no expiration for workflow execution history (effectively an infinite retention period).

Response

data RegisterDomainResponse

Response constructor

mkRegisterDomainResponse :: RegisterDomainResponse

Smart constructor for the minimum required parameters to construct a valid RegisterDomainResponse response.

This constructor is provided for convenience and testing purposes.