Network.AWS.SWF.RegisterWorkflowType

Description

Registers a new workflow type and its configuration settings in the specified domain. The retention period for the workflow history is set by the RegisterDomain action. If the type already exists, then a TypeAlreadyExists fault is returned. You cannot change the configuration settings of a workflow type once it is registered and it must be registered as a new version. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. defaultTaskList.name: String constraint. The key is swf:defaultTaskList.name. name: String constraint. The key is swf:name. version: String constraint. The key is swf:version. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_RegisterWorkflowType.html

Request

data RegisterWorkflowType

Request constructor

registerWorkflowType

Arguments

:: Text
rwtDomain

-> Text
rwtName

-> Text
rwtVersion

-> RegisterWorkflowType

RegisterWorkflowType constructor.

The fields accessible through corresponding lenses are:

rwtDefaultChildPolicy :: Maybe ChildPolicy

rwtDefaultExecutionStartToCloseTimeout :: Maybe Text

rwtDefaultTaskList :: Maybe TaskList

rwtDefaultTaskStartToCloseTimeout :: Maybe Text

rwtDescription :: Maybe Text

rwtDomain :: Text

rwtName :: Text

rwtVersion :: Text

Request lenses

rwtDefaultChildPolicy :: Lens' RegisterWorkflowType (Maybe ChildPolicy)

If set, specifies the default policy to use for the child workflow executions when a workflow execution of this type is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision. The supported child policies are: TERMINATE: the child executions will be terminated. REQUEST_CANCEL: a request to cancel will be attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. ABANDON: no action will be taken. The child executions will continue to run.

rwtDefaultExecutionStartToCloseTimeout :: Lens' RegisterWorkflowType (Maybe Text)

If set, specifies the default maximum duration for executions of this workflow type. You can override this default when starting an execution through the StartWorkflowExecution Action or StartChildWorkflowExecution Decision. The duration is specified in seconds. The valid values are integers greater than or equal to 0. Unlike some of the other timeout parameters in Amazon SWF, you cannot specify a value of NONE for defaultExecutionStartToCloseTimeout; there is a one-year max limit on the time that a workflow execution can run. Exceeding this limit will always cause the workflow execution to time out.

rwtDefaultTaskList :: Lens' RegisterWorkflowType (Maybe TaskList)

If set, specifies the default task list to use for scheduling decision tasks for executions of this workflow type. This default is used only if a task list is not provided when starting the execution through the StartWorkflowExecution Action or StartChildWorkflowExecution Decision.

rwtDefaultTaskStartToCloseTimeout :: Lens' RegisterWorkflowType (Maybe Text)

If set, specifies the default maximum duration of decision tasks for this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision. The valid values are integers greater than or equal to 0. An integer value can be used to specify the duration in seconds while NONE can be used to specify unlimited duration.

rwtDescription :: Lens' RegisterWorkflowType (Maybe Text)

Textual description of the workflow type.

rwtDomain :: Lens' RegisterWorkflowType Text

The name of the domain in which to register the workflow type.

rwtName :: Lens' RegisterWorkflowType Text

The name of the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f - u009f). Also, it must not contain the literal string &quot;arn&quot;.

rwtVersion :: Lens' RegisterWorkflowType Text

The version of the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f - u009f). Also, it must not contain the literal string &quot;arn&quot;.

Response

data RegisterWorkflowTypeResponse

Response constructor

registerWorkflowTypeResponse :: RegisterWorkflowTypeResponse

RegisterWorkflowTypeResponse constructor.