Network.AWS.SWF.RequestCancelWorkflowExecution

Description

Records a WorkflowExecutionCancelRequested event in the currently running workflow execution identified by the given domain, workflowId, and runId. This logically requests the cancellation of the workflow execution as a whole. It is up to the decider to take appropriate actions when it receives an execution history with this event.

If the runId is not specified, the WorkflowExecutionCancelRequested event is recorded in the history of the current open workflow execution with the specified workflowId in the domain. Because this action allows the workflow to properly clean up and gracefully close, it should be used instead of TerminateWorkflowExecution when possible. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_RequestCancelWorkflowExecution.html

Request

data RequestCancelWorkflowExecution

Request constructor

requestCancelWorkflowExecution

Arguments

:: Text
rcweDomain

-> Text
rcweWorkflowId

-> RequestCancelWorkflowExecution

RequestCancelWorkflowExecution constructor.

The fields accessible through corresponding lenses are:

rcweDomain :: Text

rcweRunId :: Maybe Text

rcweWorkflowId :: Text

Request lenses

rcweDomain :: Lens' RequestCancelWorkflowExecution Text

The name of the domain containing the workflow execution to cancel.

rcweRunId :: Lens' RequestCancelWorkflowExecution (Maybe Text)

The runId of the workflow execution to cancel.

rcweWorkflowId :: Lens' RequestCancelWorkflowExecution Text

The workflowId of the workflow execution to cancel.

Response

data RequestCancelWorkflowExecutionResponse

Response constructor

requestCancelWorkflowExecutionResponse :: RequestCancelWorkflowExecutionResponse

RequestCancelWorkflowExecutionResponse constructor.