Network.AWS.SWF.RespondDecisionTaskCompleted

Description

Used by deciders to tell the service that the DecisionTask identified by the taskToken has successfully completed. The decisions argument specifies the list of decisions made while processing the task.

A DecisionTaskCompleted event is added to the workflow history. The executionContext specified is attached to the event in the workflow execution history.

Access Control

If an IAM policy grants permission to use RespondDecisionTaskCompleted, it can express permissions for the list of decisions in the decisions parameter. Each of the decisions has one or more parameters, much like a regular API call. To allow for policies to be as readable as possible, you can express permissions on decisions as if they were actual API calls, including applying conditions to some parameters. For more information, see Using IAM to ManageAccess to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_RespondDecisionTaskCompleted.html

Request

data RespondDecisionTaskCompleted

Request constructor

respondDecisionTaskCompleted

Arguments

:: Text
rdtcTaskToken

-> RespondDecisionTaskCompleted

RespondDecisionTaskCompleted constructor.

The fields accessible through corresponding lenses are:

rdtcDecisions :: [Decision]

rdtcExecutionContext :: Maybe Text

rdtcTaskToken :: Text

Request lenses

rdtcDecisions :: Lens' RespondDecisionTaskCompleted [Decision]

The list of decisions (possibly empty) made by the decider while processing this decision task. See the docs for the Decision structure for details.

rdtcExecutionContext :: Lens' RespondDecisionTaskCompleted (Maybe Text)

User defined context to add to workflow execution.

rdtcTaskToken :: Lens' RespondDecisionTaskCompleted Text

The taskToken from the DecisionTask.

taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

Response

data RespondDecisionTaskCompletedResponse

Response constructor

respondDecisionTaskCompletedResponse :: RespondDecisionTaskCompletedResponse

RespondDecisionTaskCompletedResponse constructor.