Network.AWS.SWF.SignalWorkflowExecution

Description

Records a WorkflowExecutionSignaled event in the workflow execution history and creates a decision task for the workflow execution identified by the given domain, workflowId and runId. The event is recorded with the specified user defined signalName and input (if provided).

If a runId is not specified, then the WorkflowExecutionSignaled event is recorded in the history of the current open workflow with the matching workflowId in the domain. If the specified workflow execution is not open, this method fails with UnknownResource. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_SignalWorkflowExecution.html

Request

data SignalWorkflowExecution

Request constructor

signalWorkflowExecution

Arguments

:: Text
sweDomain

-> Text
sweWorkflowId

-> Text
sweSignalName

-> SignalWorkflowExecution

SignalWorkflowExecution constructor.

The fields accessible through corresponding lenses are:

sweDomain :: Text

sweInput :: Maybe Text

sweRunId :: Maybe Text

sweSignalName :: Text

sweWorkflowId :: Text

Request lenses

sweDomain :: Lens' SignalWorkflowExecution Text

The name of the domain containing the workflow execution to signal.

sweInput :: Lens' SignalWorkflowExecution (Maybe Text)

Data to attach to the WorkflowExecutionSignaled event in the target workflow execution's history.

sweRunId :: Lens' SignalWorkflowExecution (Maybe Text)

The runId of the workflow execution to signal.

sweSignalName :: Lens' SignalWorkflowExecution Text

The name of the signal. This name must be meaningful to the target workflow.

sweWorkflowId :: Lens' SignalWorkflowExecution Text

The workflowId of the workflow execution to signal.

Response

data SignalWorkflowExecutionResponse

Response constructor

signalWorkflowExecutionResponse :: SignalWorkflowExecutionResponse

SignalWorkflowExecutionResponse constructor.