Network.AWS.SWF.StartWorkflowExecution

Description

Starts an execution of the workflow type in the specified domain using the provided workflowId and input data. This action returns the newly started workflow execution. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. tagList.member.0: The key is swf:tagList.member.0. tagList.member.1: The key is swf:tagList.member.1. tagList.member.2: The key is swf:tagList.member.2. tagList.member.3: The key is swf:tagList.member.3. tagList.member.4: The key is swf:tagList.member.4. taskList: String constraint. The key is swf:taskList.name. name: String constraint. The key is swf:workflowType.name. version: String constraint. The key is swf:workflowType.version. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_StartWorkflowExecution.html

Request

data StartWorkflowExecution

Request constructor

startWorkflowExecution

Arguments

:: Text
swe1Domain

-> Text
swe1WorkflowId

-> WorkflowType
swe1WorkflowType

-> StartWorkflowExecution

StartWorkflowExecution constructor.

The fields accessible through corresponding lenses are:

swe1ChildPolicy :: Maybe ChildPolicy

swe1Domain :: Text

swe1ExecutionStartToCloseTimeout :: Maybe Text

swe1Input :: Maybe Text

swe1TagList :: [Text]

swe1TaskList :: Maybe TaskList

swe1TaskStartToCloseTimeout :: Maybe Text

swe1WorkflowId :: Text

swe1WorkflowType :: WorkflowType

Request lenses

swe1ChildPolicy :: Lens' StartWorkflowExecution (Maybe ChildPolicy)

If set, specifies the policy to use for the child workflow executions of this workflow execution if it is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This policy overrides the default child policy specified when registering the workflow type using RegisterWorkflowType. The supported child policies are: TERMINATE: the child executions will be terminated. REQUEST_CANCEL: a request to cancel will be attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. ABANDON: no action will be taken. The child executions will continue to run.

swe1Domain :: Lens' StartWorkflowExecution Text

The name of the domain in which the workflow execution is created.

swe1ExecutionStartToCloseTimeout :: Lens' StartWorkflowExecution (Maybe Text)

The total duration for this workflow execution. This overrides the defaultExecutionStartToCloseTimeout specified when registering the workflow type. The duration is specified in seconds. The valid values are integers greater than or equal to 0. Exceeding this limit will cause the workflow execution to time out. Unlike some of the other timeout parameters in Amazon SWF, you cannot specify a value of NONE for this timeout; there is a one-year max limit on the time that a workflow execution can run.

swe1Input :: Lens' StartWorkflowExecution (Maybe Text)

The input for the workflow execution. This is a free form string which should be meaningful to the workflow you are starting. This input is made available to the new workflow execution in the WorkflowExecutionStarted history event.

swe1TagList :: Lens' StartWorkflowExecution [Text]

The list of tags to associate with the workflow execution. You can specify a maximum of 5 tags. You can list workflow executions with a specific tag by calling ListOpenWorkflowExecutions or ListClosedWorkflowExecutions and specifying a TagFilter.

swe1TaskList :: Lens' StartWorkflowExecution (Maybe TaskList)

The task list to use for the decision tasks generated for this workflow execution. This overrides the defaultTaskList specified when registering the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f - u009f). Also, it must not contain the literal string &quot;arn&quot;.

swe1TaskStartToCloseTimeout :: Lens' StartWorkflowExecution (Maybe Text)

Specifies the maximum duration of decision tasks for this workflow execution. This parameter overrides the defaultTaskStartToCloseTimout specified when registering the workflow type using RegisterWorkflowType. The valid values are integers greater than or equal to 0. An integer value can be used to specify the duration in seconds while NONE can be used to specify unlimited duration.

swe1WorkflowId :: Lens' StartWorkflowExecution Text

The user defined identifier associated with the workflow execution. You can use this to associate a custom identifier with the workflow execution. You may specify the same identifier if a workflow execution is logically a restart of a previous execution. You cannot have two open workflow executions with the same workflowId at the same time. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f - u009f). Also, it must not contain the literal string &quot;arn&quot;.

swe1WorkflowType :: Lens' StartWorkflowExecution WorkflowType

The type of the workflow to start.

Response

data StartWorkflowExecutionResponse

Response constructor

startWorkflowExecutionResponse :: StartWorkflowExecutionResponse

StartWorkflowExecutionResponse constructor.

The fields accessible through corresponding lenses are:

swerRunId :: Maybe Text

Response lenses

swerRunId :: Lens' StartWorkflowExecutionResponse (Maybe Text)

The runId of a workflow execution. This Id is generated by the service and can be used to uniquely identify the workflow execution within a domain.