Network.AWS.SWF.TerminateWorkflowExecution

Description

Records a WorkflowExecutionTerminated event and forces closure of the workflow execution identified by the given domain, runId, and workflowId. The child policy, registered with the workflow type or specified when starting this execution, is applied to any open child workflow executions of this workflow execution. If the identified workflow execution was in progress, it is terminated immediately. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_TerminateWorkflowExecution.html

Request

data TerminateWorkflowExecution

Request constructor

terminateWorkflowExecution

Arguments

:: Text
tweDomain

-> Text
tweWorkflowId

-> TerminateWorkflowExecution

TerminateWorkflowExecution constructor.

The fields accessible through corresponding lenses are:

tweChildPolicy :: Maybe Text

tweDetails :: Maybe Text

tweDomain :: Text

tweReason :: Maybe Text

tweRunId :: Maybe Text

tweWorkflowId :: Text

Request lenses

tweChildPolicy :: Lens' TerminateWorkflowExecution (Maybe Text)

If set, specifies the policy to use for the child workflow executions of the workflow execution being terminated. This policy overrides the child policy specified for the workflow execution at registration time or when starting the execution. The supported child policies are: TERMINATE: the child executions will be terminated. REQUEST_CANCEL: a request to cancel will be attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. ABANDON: no action will be taken. The child executions will continue to run.

tweDetails :: Lens' TerminateWorkflowExecution (Maybe Text)

Optional details for terminating the workflow execution.

tweDomain :: Lens' TerminateWorkflowExecution Text

The domain of the workflow execution to terminate.

tweReason :: Lens' TerminateWorkflowExecution (Maybe Text)

An optional descriptive reason for terminating the workflow execution.

tweRunId :: Lens' TerminateWorkflowExecution (Maybe Text)

The runId of the workflow execution to terminate.

tweWorkflowId :: Lens' TerminateWorkflowExecution Text

The workflowId of the workflow execution to terminate.

Response

data TerminateWorkflowExecutionResponse

Response constructor

terminateWorkflowExecutionResponse :: TerminateWorkflowExecutionResponse

TerminateWorkflowExecutionResponse constructor.