Network.AWS.SWF.V2012_01_25.DeprecateActivityType

Description

Deprecates the specified activity type. After an activity type has been deprecated, you cannot create new tasks of that activity type. Tasks of this type that were scheduled before the type was deprecated will continue to run. This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. activityType.name: String constraint. The key is swf:activityType.name. activityType.version: String constraint. The key is swf:activityType.version. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows. DeprecateActivityType Example POST HTTP1.1 Host: swf.us-east-1.amazonaws.com User-Agent: Mozilla/5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.2.25) Gecko/20111212 Firefox3.6.25 ( .NET CLR 3.5.30729; .NET4.0E) Accept: applicationjson, textjavascript, ** Accept-Language: en-us,en;q=0.5 Accept-Encoding: gzip,deflate Accept-Charset: ISO-8859-1,utf-8;q=0.7,*;q=0.7 Keep-Alive: 115 Connection: keep-alive Content-Type: application/x-amz-json-1.0 X-Requested-With: XMLHttpRequest X-Amz-Date: Mon, 16 Jan 2012 05:01:06 GMT X-Amz-Target: SimpleWorkflowService.DeprecateActivityType Content-Encoding: amz-1.0 X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE,Algorithm=HmacSHA256,SignedHeaders=Host;X-Amz-Date;X-Amz-Target;Content-Encoding,Signature=iX/mNMtNH6IaSNwfZq9hHOhDlLnp7buuj9tO93kRIrQ= Referer: http:/swf.us-east-1.amazonaws.comexplorer/index.html Content-Length: 95 Pragma: no-cache Cache-Control: no-cache {"domain": "867530901", "activityType": {"name": "activityVerify", "version": "1.0"} } HTTP1.1 200 OK Content-Length: 0 Content-Type: applicationjson x-amzn-RequestId: 191ee17e-3fff-11e1-a23a-99d60383ae71.

Documentation

data DeprecateActivityType

Constructors

DeprecateActivityType

Fields

_datjActivityType :: ActivityType

The activity type to deprecate.

_datjDomain :: Text

The name of the domain in which the activity type is registered.

datjDomain :: Lens' DeprecateActivityType Text

datjActivityType :: Lens' DeprecateActivityType ActivityType

data DeprecateActivityTypeResponse

Constructors

DeprecateActivityTypeResponse