Network.AWS.SWF.V2012_01_25.RequestCancelWorkflowExecution

Description

Records a WorkflowExecutionCancelRequested event in the currently running workflow execution identified by the given domain, workflowId, and runId. This logically requests the cancellation of the workflow execution as a whole. It is up to the decider to take appropriate actions when it receives an execution history with this event. If the runId is not specified, the WorkflowExecutionCancelRequested event is recorded in the history of the current open workflow execution with the specified workflowId in the domain. Because this action allows the workflow to properly clean up and gracefully close, it should be used instead of TerminateWorkflowExecution when possible. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows. RequestCancelWorkflowExecution Example POST HTTP1.1 Host: swf.us-east-1.amazonaws.com User-Agent: Mozilla5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.2.25) Gecko20111212 Firefox3.6.25 ( .NET CLR 3.5.30729; .NET4.0E) Accept: applicationjson, textjavascript, ** Accept-Language: en-us,en;q=0.5 Accept-Encoding: gzip,deflate Accept-Charset: ISO-8859-1,utf-8;q=0.7,*;q=0.7 Keep-Alive: 115 Connection: keep-alive Content-Type: application/x-amz-json-1.0 X-Requested-With: XMLHttpRequest X-Amz-Date: Mon, 16 Jan 2012 04:49:06 GMT X-Amz-Target: SimpleWorkflowService.RequestCancelWorkflowExecution Content-Encoding: amz-1.0 X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE,Algorithm=HmacSHA256,SignedHeaders=Host;X-Amz-Date;X-Amz-Target;Content-Encoding,Signature=xODwV3kbpJbWVa6bQiV2zQAw9euGI3uXI82urc+bVeo= Referer: http:/swf.us-east-1.amazonaws.comexplorer/index.html Content-Length: 106 Pragma: no-cache Cache-Control: no-cache {"domain": "867530901", "workflowId": "20110927-T-1", "runId": "94861fda-a714-4126-95d7-55ba847da8ab"} HTTP/1.1 200 OK Content-Length: 0 Content-Type: application/json x-amzn-RequestId: 6bd0627e-3ffd-11e1-9b11-7182192d0b57.

Documentation

requestCancelWorkflowExecution

Arguments

:: Text
_rcweiDomain

-> Text
_rcweiWorkflowId

-> RequestCancelWorkflowExecution

Minimum specification for a RequestCancelWorkflowExecution request.

data RequestCancelWorkflowExecution

Constructors

RequestCancelWorkflowExecution

Fields

_rcweiDomain :: Text

The name of the domain containing the workflow execution to cancel.

_rcweiWorkflowId :: Text

The workflowId of the workflow execution to cancel.

_rcweiRunId :: Maybe Text

The runId of the workflow execution to cancel.

rcweiWorkflowId :: Lens' RequestCancelWorkflowExecution Text

rcweiRunId :: Lens' RequestCancelWorkflowExecution (Maybe Text)

rcweiDomain :: Lens' RequestCancelWorkflowExecution Text

data RequestCancelWorkflowExecutionResponse

Constructors

RequestCancelWorkflowExecutionResponse