Network.AWS.SWF.V2012_01_25.RespondActivityTaskFailed

Description

Used by workers to tell the service that the ActivityTask identified by the taskToken has failed with reason (if specified). The reason and details appear in the ActivityTaskFailed event added to the workflow history. A task is considered open from the time that it is scheduled until it is closed. Therefore a task is reported as open while a worker is processing it. A task is closed after it has been specified in a call to RespondActivityTaskCompleted, RespondActivityTaskCanceled, RespondActivityTaskFailed, or the task has timed out. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows. RespondActivityTaskFailed Example POST HTTP1.1 Host: swf.us-east-1.amazonaws.com User-Agent: Mozilla/5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.2.25) Gecko20111212 Firefox3.6.25 ( .NET CLR 3.5.30729; .NET4.0E) Accept: applicationjson, textjavascript, */* Accept-Language: en-us,en;q=0.5 Accept-Encoding: gzip,deflate Accept-Charset: ISO-8859-1,utf-8;q=0.7,*;q=0.7 Keep-Alive: 115 Connection: keep-alive Content-Type: application/x-amz-json-1.0 X-Requested-With: XMLHttpRequest X-Amz-Date: Mon, 16 Jan 2012 04:17:24 GMT X-Amz-Target: SimpleWorkflowService.RespondActivityTaskFailed Content-Encoding: amz-1.0 X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE,Algorithm=HmacSHA256,SignedHeaders=Host;X-Amz-Date;X-Amz-Target;Content-Encoding,Signature=JC+udsmFEq8qca2WFs5kfp2eAEONc70IqFgHErhpc= Referer: http:/swf.us-east-1.amazonaws.comexplorer/index.html Content-Length: 682 Pragma: no-cache Cache-Control: no-cache {"taskToken": "AAAAKgAAAAEAAAAAAAAAAdG7j7YFEl9pfKdXRL3Cy3Q3c1Z8QwdOSX53bKiUV6MMGXvf3Lrinmmzj1HFFl5lcwHzEFxLbMaSZlMtRFJPumHXAnUqlYjZLODhrBqsIzDQFKcbCFMq7y4jm0EFzsV2Suv8iuobcZidU8qjd9uG82zumG2xz1Z4IbOFwOTlpj2++5YVH4ftyycIcjlDw58r0O1vAo4PEondkqjyn+YxBxyZLy1z1fvMi0zeO8Lh16w96y6v+KdVcECoez1Og8sROaXG0l8ptW5YR733LIuUBK4sxWa12egF5i4e8AV8JloojOaq0jy4iFsIscRazOSQErjo15Guz89BK2XW911P3I+X7nJjH0wwW55XGCs0jezvsEC8M6D9Ob7CgWr6RrnK3g1AKemcby2XqgQRN52DMIYxzV+lMSQBYKOqtkLoMY0NKeuRVwm9f1zCY00v6kxqK9m2zFvaxqlJ5JVCWMNWEWJfQZVtC3GzMWmzeCt7Auq8A5CaqDKyOhTIhY/Go00iiDA6ecP8taTYiVzb8VR5xEiQ1uCxnECkwW", "reason": "could not verify customer credit card", "details": "card number invalid"} HTTP1.1 200 OK Content-Length: 0 Content-Type: applicationjson x-amzn-RequestId: feadaedd-3ff8-11e1-9e8f-57bb03e21482.

Request

data RespondActivityTaskFailed

Request constructor

mkRespondActivityTaskFailed

Arguments

:: Text
ratfTaskToken

-> RespondActivityTaskFailed

Smart constructor for the minimum required parameters to construct a valid RespondActivityTaskFailed request.

The fields accessible through corresponding lenses are:

TaskToken :: Text

Reason :: Maybe Text

Details :: Maybe Text

Request lenses

ratfTaskToken :: Lens' RespondActivityTaskFailed Text

The taskToken of the ActivityTask. The taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

ratfReason :: Lens' RespondActivityTaskFailed (Maybe Text)

Description of the error that may assist in diagnostics.

ratfDetails :: Lens' RespondActivityTaskFailed (Maybe Text)

Optional detailed information about the failure.

Response

data RespondActivityTaskFailedResponse

Response constructor

mkRespondActivityTaskFailedResponse :: RespondActivityTaskFailedResponse

Smart constructor for the minimum required parameters to construct a valid RespondActivityTaskFailedResponse response.

This constructor is provided for convenience and testing purposes.