Network.AWS.SWF.CountOpenWorkflowExecutions

Description

Returns the number of open workflow executions within the given domain that meet the specified filtering criteria.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagFilter.tag : String constraint. The key is swf:tagFilter.tag .

typeFilter.name : String constraint. The key is swf:typeFilter.name .

typeFilter.version : String constraint. The key is swf:typeFilter.version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

countOpenWorkflowExecutions

Arguments

:: Text
coweDomain

-> ExecutionTimeFilter
coweStartTimeFilter

-> CountOpenWorkflowExecutions

Creates a value of CountOpenWorkflowExecutions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

coweExecutionFilter - If specified, only workflow executions matching the WorkflowId in the filter are counted.

coweTypeFilter - Specifies the type of the workflow executions to be counted.

coweTagFilter - If specified, only executions that have a tag that matches the filter are counted.

coweDomain - The name of the domain containing the workflow executions to count.

coweStartTimeFilter - Specifies the start time criteria that workflow executions must meet in order to be counted.

data CountOpenWorkflowExecutions

See: countOpenWorkflowExecutions smart constructor.

Request Lenses

coweExecutionFilter :: Lens' CountOpenWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the WorkflowId in the filter are counted.

coweTypeFilter :: Lens' CountOpenWorkflowExecutions (Maybe WorkflowTypeFilter)

Specifies the type of the workflow executions to be counted.

coweTagFilter :: Lens' CountOpenWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have a tag that matches the filter are counted.

coweDomain :: Lens' CountOpenWorkflowExecutions Text

The name of the domain containing the workflow executions to count.

coweStartTimeFilter :: Lens' CountOpenWorkflowExecutions ExecutionTimeFilter

Specifies the start time criteria that workflow executions must meet in order to be counted.

Destructuring the Response

workflowExecutionCount

Arguments

:: Natural
wecCount

-> WorkflowExecutionCount

Creates a value of WorkflowExecutionCount with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

wecTruncated - If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

wecCount - The number of workflow executions.

data WorkflowExecutionCount

Contains the count of workflow executions returned from CountOpenWorkflowExecutions or CountClosedWorkflowExecutions

See: workflowExecutionCount smart constructor.

Response Lenses

wecTruncated :: Lens' WorkflowExecutionCount (Maybe Bool)

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

wecCount :: Lens' WorkflowExecutionCount Natural

The number of workflow executions.