Network.AWS.SWF.CountPendingActivityTasks

Description

Returns the estimated number of activity tasks in the specified task list. The count returned is an approximation and isn't guaranteed to be exact. If you specify a task list that no activity task was ever scheduled in then 0 is returned.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the taskList.name parameter by using a Condition element with the swf:taskList.name key to allow the action to access only certain task lists.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

countPendingActivityTasks

Arguments

:: Text
cpatDomain

-> TaskList
cpatTaskList

-> CountPendingActivityTasks

Creates a value of CountPendingActivityTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpatDomain - The name of the domain that contains the task list.

cpatTaskList - The name of the task list.

data CountPendingActivityTasks

See: countPendingActivityTasks smart constructor.

Request Lenses

cpatDomain :: Lens' CountPendingActivityTasks Text

The name of the domain that contains the task list.

cpatTaskList :: Lens' CountPendingActivityTasks TaskList

The name of the task list.

Destructuring the Response

pendingTaskCount

Arguments

:: Natural
ptcCount

-> PendingTaskCount

Creates a value of PendingTaskCount with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ptcTruncated - If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

ptcCount - The number of tasks in the task list.

data PendingTaskCount

Contains the count of tasks in a task list.

See: pendingTaskCount smart constructor.

Response Lenses

ptcTruncated :: Lens' PendingTaskCount (Maybe Bool)

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

ptcCount :: Lens' PendingTaskCount Natural

The number of tasks in the task list.