Network.AWS.SWF.CountPendingDecisionTasks

Description

Returns the estimated number of decision tasks in the specified task list. The count returned is an approximation and is not guaranteed to be exact. If you specify a task list that no decision task was ever scheduled in then 0 will be returned.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the 'taskList.name' parameter by using a Condition element with the 'swf:taskList.name' key to allow the action to access only certain task lists. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_CountPendingDecisionTasks.html

Request

data CountPendingDecisionTasks

Request constructor

countPendingDecisionTasks

Arguments

:: Text
cpdtDomain

-> TaskList
cpdtTaskList

-> CountPendingDecisionTasks

CountPendingDecisionTasks constructor.

The fields accessible through corresponding lenses are:

cpdtDomain :: Text

cpdtTaskList :: TaskList

Request lenses

cpdtDomain :: Lens' CountPendingDecisionTasks Text

The name of the domain that contains the task list.

cpdtTaskList :: Lens' CountPendingDecisionTasks TaskList

The name of the task list.

Response

data CountPendingDecisionTasksResponse

Response constructor

countPendingDecisionTasksResponse

Arguments

:: Natural
cpdtrCount

-> CountPendingDecisionTasksResponse

CountPendingDecisionTasksResponse constructor.

The fields accessible through corresponding lenses are:

cpdtrCount :: Natural

cpdtrTruncated :: Maybe Bool

Response lenses

cpdtrCount :: Lens' CountPendingDecisionTasksResponse Natural

The number of tasks in the task list.

cpdtrTruncated :: Lens' CountPendingDecisionTasksResponse (Maybe Bool)

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.