Network.AWS.SWF.DeprecateActivityType

Description

Deprecates the specified activity type . After an activity type has been deprecated, you cannot create new tasks of that activity type. Tasks of this type that were scheduled before the type was deprecated continue to run.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

activityType.name : String constraint. The key is swf:activityType.name .

activityType.version : String constraint. The key is swf:activityType.version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

deprecateActivityType

Arguments

:: Text
depDomain

-> ActivityType
depActivityType

-> DeprecateActivityType

Creates a value of DeprecateActivityType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

depDomain - The name of the domain in which the activity type is registered.

depActivityType - The activity type to deprecate.

data DeprecateActivityType

See: deprecateActivityType smart constructor.

Request Lenses

depDomain :: Lens' DeprecateActivityType Text

The name of the domain in which the activity type is registered.

depActivityType :: Lens' DeprecateActivityType ActivityType

The activity type to deprecate.

Destructuring the Response

deprecateActivityTypeResponse :: DeprecateActivityTypeResponse

Creates a value of DeprecateActivityTypeResponse with the minimum fields required to make a request.

data DeprecateActivityTypeResponse

See: deprecateActivityTypeResponse smart constructor.