Network.AWS.SWF.DeprecateDomain

Description

Deprecates the specified domain. After a domain has been deprecated it cannot be used to create new workflow executions or register new types. However, you can still use visibility actions on this domain. Deprecating a domain also deprecates all activity and workflow types registered in the domain. Executions that were started before the domain was deprecated continues to run.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

deprecateDomain

Arguments

:: Text
dName

-> DeprecateDomain

Creates a value of DeprecateDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dName - The name of the domain to deprecate.

data DeprecateDomain

See: deprecateDomain smart constructor.

Request Lenses

dName :: Lens' DeprecateDomain Text

The name of the domain to deprecate.

Destructuring the Response

deprecateDomainResponse :: DeprecateDomainResponse

Creates a value of DeprecateDomainResponse with the minimum fields required to make a request.

data DeprecateDomainResponse

See: deprecateDomainResponse smart constructor.