Network.AWS.SWF.DescribeActivityType

Description

Returns information about the specified activity type. This includes configuration settings provided when the type was registered and other general information about the type.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

activityType.name : String constraint. The key is swf:activityType.name .

activityType.version : String constraint. The key is swf:activityType.version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

describeActivityType

Arguments

:: Text
datDomain

-> ActivityType
datActivityType

-> DescribeActivityType

Creates a value of DescribeActivityType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

datDomain - The name of the domain in which the activity type is registered.

datActivityType - The activity type to get information about. Activity types are identified by the name and version that were supplied when the activity was registered.

data DescribeActivityType

See: describeActivityType smart constructor.

Request Lenses

datDomain :: Lens' DescribeActivityType Text

The name of the domain in which the activity type is registered.

datActivityType :: Lens' DescribeActivityType ActivityType

The activity type to get information about. Activity types are identified by the name and version that were supplied when the activity was registered.

Destructuring the Response

describeActivityTypeResponse

Arguments

:: Int
datrsResponseStatus

-> ActivityTypeInfo
datrsTypeInfo

-> ActivityTypeConfiguration
datrsConfiguration

-> DescribeActivityTypeResponse

Creates a value of DescribeActivityTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

datrsResponseStatus - -- | The response status code.

datrsTypeInfo - General information about the activity type. The status of activity type (returned in the ActivityTypeInfo structure) can be one of the following. * REGISTERED – The type is registered and available. Workers supporting this type should be running. * DEPRECATED – The type was deprecated using DeprecateActivityType , but is still in use. You should keep workers supporting this type running. You cannot create new tasks of this type.

datrsConfiguration - The configuration settings registered with the activity type.

data DescribeActivityTypeResponse

Detailed information about an activity type.

See: describeActivityTypeResponse smart constructor.

Response Lenses

datrsResponseStatus :: Lens' DescribeActivityTypeResponse Int

- | The response status code.

datrsTypeInfo :: Lens' DescribeActivityTypeResponse ActivityTypeInfo

General information about the activity type. The status of activity type (returned in the ActivityTypeInfo structure) can be one of the following. * REGISTERED – The type is registered and available. Workers supporting this type should be running. * DEPRECATED – The type was deprecated using DeprecateActivityType , but is still in use. You should keep workers supporting this type running. You cannot create new tasks of this type.

datrsConfiguration :: Lens' DescribeActivityTypeResponse ActivityTypeConfiguration

The configuration settings registered with the activity type.