Network.AWS.SWF.DescribeWorkflowExecution

Description

Returns information about the specified workflow execution including its type and some statistics.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

describeWorkflowExecution

Arguments

:: Text
dweDomain

-> WorkflowExecution
dweExecution

-> DescribeWorkflowExecution

Creates a value of DescribeWorkflowExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dweDomain - The name of the domain containing the workflow execution.

dweExecution - The workflow execution to describe.

data DescribeWorkflowExecution

See: describeWorkflowExecution smart constructor.

Request Lenses

dweDomain :: Lens' DescribeWorkflowExecution Text

The name of the domain containing the workflow execution.

dweExecution :: Lens' DescribeWorkflowExecution WorkflowExecution

The workflow execution to describe.

Destructuring the Response

describeWorkflowExecutionResponse

Arguments

:: Int
dwersResponseStatus

-> WorkflowExecutionInfo
dwersExecutionInfo

-> WorkflowExecutionConfiguration
dwersExecutionConfiguration

-> WorkflowExecutionOpenCounts
dwersOpenCounts

-> DescribeWorkflowExecutionResponse

Creates a value of DescribeWorkflowExecutionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwersLatestActivityTaskTimestamp - The time when the last activity task was scheduled for this workflow execution. You can use this information to determine if the workflow has not made progress for an unusually long period of time and might require a corrective action.

dwersLatestExecutionContext - The latest executionContext provided by the decider for this workflow execution. A decider can provide an executionContext (a free-form string) when closing a decision task using RespondDecisionTaskCompleted .

dwersResponseStatus - -- | The response status code.

dwersExecutionInfo - Information about the workflow execution.

dwersExecutionConfiguration - The configuration settings for this workflow execution including timeout values, tasklist etc.

dwersOpenCounts - The number of tasks for this workflow execution. This includes open and closed tasks of all types.

data DescribeWorkflowExecutionResponse

Contains details about a workflow execution.

See: describeWorkflowExecutionResponse smart constructor.

Response Lenses

dwersLatestActivityTaskTimestamp :: Lens' DescribeWorkflowExecutionResponse (Maybe UTCTime)

The time when the last activity task was scheduled for this workflow execution. You can use this information to determine if the workflow has not made progress for an unusually long period of time and might require a corrective action.

dwersLatestExecutionContext :: Lens' DescribeWorkflowExecutionResponse (Maybe Text)

The latest executionContext provided by the decider for this workflow execution. A decider can provide an executionContext (a free-form string) when closing a decision task using RespondDecisionTaskCompleted .

dwersResponseStatus :: Lens' DescribeWorkflowExecutionResponse Int

- | The response status code.

dwersExecutionInfo :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionInfo

Information about the workflow execution.

dwersExecutionConfiguration :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionConfiguration

The configuration settings for this workflow execution including timeout values, tasklist etc.

dwersOpenCounts :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionOpenCounts

The number of tasks for this workflow execution. This includes open and closed tasks of all types.