Network.AWS.SWF.DescribeWorkflowType

Description

Returns information about the specified workflow type . This includes configuration settings specified when the type was registered and other information such as creation date, current status, etc.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

workflowType.name : String constraint. The key is swf:workflowType.name .

workflowType.version : String constraint. The key is swf:workflowType.version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

describeWorkflowType

Arguments

:: Text
dwtDomain

-> WorkflowType
dwtWorkflowType

-> DescribeWorkflowType

Creates a value of DescribeWorkflowType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwtDomain - The name of the domain in which this workflow type is registered.

dwtWorkflowType - The workflow type to describe.

data DescribeWorkflowType

See: describeWorkflowType smart constructor.

Request Lenses

dwtDomain :: Lens' DescribeWorkflowType Text

The name of the domain in which this workflow type is registered.

dwtWorkflowType :: Lens' DescribeWorkflowType WorkflowType

The workflow type to describe.

Destructuring the Response

describeWorkflowTypeResponse

Arguments

:: Int
dwtrsResponseStatus

-> WorkflowTypeInfo
dwtrsTypeInfo

-> WorkflowTypeConfiguration
dwtrsConfiguration

-> DescribeWorkflowTypeResponse

Creates a value of DescribeWorkflowTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwtrsResponseStatus - -- | The response status code.

dwtrsTypeInfo - General information about the workflow type. The status of the workflow type (returned in the WorkflowTypeInfo structure) can be one of the following. * REGISTERED – The type is registered and available. Workers supporting this type should be running. * DEPRECATED – The type was deprecated using DeprecateWorkflowType , but is still in use. You should keep workers supporting this type running. You cannot create new workflow executions of this type.

dwtrsConfiguration - Configuration settings of the workflow type registered through RegisterWorkflowType

data DescribeWorkflowTypeResponse

Contains details about a workflow type.

See: describeWorkflowTypeResponse smart constructor.

Response Lenses

dwtrsResponseStatus :: Lens' DescribeWorkflowTypeResponse Int

- | The response status code.

dwtrsTypeInfo :: Lens' DescribeWorkflowTypeResponse WorkflowTypeInfo

General information about the workflow type. The status of the workflow type (returned in the WorkflowTypeInfo structure) can be one of the following. * REGISTERED – The type is registered and available. Workers supporting this type should be running. * DEPRECATED – The type was deprecated using DeprecateWorkflowType , but is still in use. You should keep workers supporting this type running. You cannot create new workflow executions of this type.

dwtrsConfiguration :: Lens' DescribeWorkflowTypeResponse WorkflowTypeConfiguration

Configuration settings of the workflow type registered through RegisterWorkflowType