Network.AWS.SWF.DescribeWorkflowType

Description

Returns information about the specified workflow type. This includes configuration settings specified when the type was registered and other information such as creation date, current status, etc.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. 'workflowType.name': String constraint. The key is 'swf:workflowType.name'. 'workflowType.version': String constraint. The key is 'swf:workflowType.version' . If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_DescribeWorkflowType.html

Request

data DescribeWorkflowType

Request constructor

describeWorkflowType

Arguments

:: Text
dwtDomain

-> WorkflowType
dwtWorkflowType

-> DescribeWorkflowType

DescribeWorkflowType constructor.

The fields accessible through corresponding lenses are:

dwtDomain :: Text

dwtWorkflowType :: WorkflowType

Request lenses

dwtDomain :: Lens' DescribeWorkflowType Text

The name of the domain in which this workflow type is registered.

dwtWorkflowType :: Lens' DescribeWorkflowType WorkflowType

The workflow type to describe.

Response

data DescribeWorkflowTypeResponse

Response constructor

describeWorkflowTypeResponse

Arguments

:: WorkflowTypeInfo
dwtrTypeInfo

-> WorkflowTypeConfiguration
dwtrConfiguration

-> DescribeWorkflowTypeResponse

DescribeWorkflowTypeResponse constructor.

The fields accessible through corresponding lenses are:

dwtrConfiguration :: WorkflowTypeConfiguration

dwtrTypeInfo :: WorkflowTypeInfo

Response lenses

dwtrConfiguration :: Lens' DescribeWorkflowTypeResponse WorkflowTypeConfiguration

Configuration settings of the workflow type registered through RegisterWorkflowType

dwtrTypeInfo :: Lens' DescribeWorkflowTypeResponse WorkflowTypeInfo

General information about the workflow type.

The status of the workflow type (returned in the WorkflowTypeInfo structure) can be one of the following.

REGISTERED: The type is registered and available. Workers supporting this type should be running. DEPRECATED: The type was deprecated using DeprecateWorkflowType, but is still in use. You should keep workers supporting this type running. You cannot create new workflow executions of this type.