Network.AWS.SWF.GetWorkflowExecutionHistory

Description

Returns the history of the specified workflow execution. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_GetWorkflowExecutionHistory.html

Request

data GetWorkflowExecutionHistory

Request constructor

getWorkflowExecutionHistory

Arguments

:: Text
gwehDomain

-> WorkflowExecution
gwehExecution

-> GetWorkflowExecutionHistory

GetWorkflowExecutionHistory constructor.

The fields accessible through corresponding lenses are:

gwehDomain :: Text

gwehExecution :: WorkflowExecution

gwehMaximumPageSize :: Maybe Natural

gwehNextPageToken :: Maybe Text

gwehReverseOrder :: Maybe Bool

Request lenses

gwehDomain :: Lens' GetWorkflowExecutionHistory Text

The name of the domain containing the workflow execution.

gwehExecution :: Lens' GetWorkflowExecutionHistory WorkflowExecution

Specifies the workflow execution for which to return the history.

gwehMaximumPageSize :: Lens' GetWorkflowExecutionHistory (Maybe Natural)

The maximum number of results that will be returned per call. nextPageToken can be used to obtain futher pages of results. The default is 100, which is the maximum allowed page size. You can, however, specify a page size smaller than 100.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

gwehNextPageToken :: Lens' GetWorkflowExecutionHistory (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

gwehReverseOrder :: Lens' GetWorkflowExecutionHistory (Maybe Bool)

When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimeStamp of the events.

Response

data GetWorkflowExecutionHistoryResponse

Response constructor

getWorkflowExecutionHistoryResponse :: GetWorkflowExecutionHistoryResponse

GetWorkflowExecutionHistoryResponse constructor.

The fields accessible through corresponding lenses are:

gwehrEvents :: [HistoryEvent]

gwehrNextPageToken :: Maybe Text

Response lenses

gwehrEvents :: Lens' GetWorkflowExecutionHistoryResponse [HistoryEvent]

The list of history events.

gwehrNextPageToken :: Lens' GetWorkflowExecutionHistoryResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.