Network.AWS.SWF.ListActivityTypes

Description

Returns information about all activities registered in the specified domain that match the specified name and registration status. The result includes information like creation date, current status of the activity, etc. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_ListActivityTypes.html

Request

data ListActivityTypes

Request constructor

listActivityTypes

Arguments

:: Text
latDomain

-> RegistrationStatus
latRegistrationStatus

-> ListActivityTypes

ListActivityTypes constructor.

The fields accessible through corresponding lenses are:

latDomain :: Text

latMaximumPageSize :: Maybe Natural

latName :: Maybe Text

latNextPageToken :: Maybe Text

latRegistrationStatus :: RegistrationStatus

latReverseOrder :: Maybe Bool

Request lenses

latDomain :: Lens' ListActivityTypes Text

The name of the domain in which the activity types have been registered.

latMaximumPageSize :: Lens' ListActivityTypes (Maybe Natural)

The maximum number of results returned in each page. The default is 100, but the caller can override this value to a page size smaller than the default. You cannot specify a page size greater than 100. Note that the number of types may be less than the maxiumum page size, in which case, the returned page will have fewer results than the maximumPageSize specified.

latName :: Lens' ListActivityTypes (Maybe Text)

If specified, only lists the activity types that have this name.

latNextPageToken :: Lens' ListActivityTypes (Maybe Text)

If on a previous call to this method a NextResultToken was returned, the results have more than one page. To get the next page of results, repeat the call with the nextPageToken and keep all other arguments unchanged.

latRegistrationStatus :: Lens' ListActivityTypes RegistrationStatus

Specifies the registration status of the activity types to list.

latReverseOrder :: Lens' ListActivityTypes (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in ascending alphabetical order of the name of the activity types.

Response

data ListActivityTypesResponse

Response constructor

listActivityTypesResponse :: ListActivityTypesResponse

ListActivityTypesResponse constructor.

The fields accessible through corresponding lenses are:

latrNextPageToken :: Maybe Text

latrTypeInfos :: [ActivityTypeInfo]

Response lenses

latrNextPageToken :: Lens' ListActivityTypesResponse (Maybe Text)

Returns a value if the results are paginated. To get the next page of results, repeat the request specifying this token and all other arguments unchanged.

latrTypeInfos :: Lens' ListActivityTypesResponse [ActivityTypeInfo]

List of activity type information.