Network.AWS.SWF.ListActivityTypes

Description

Returns information about all activities registered in the specified domain that match the specified name and registration status. The result includes information like creation date, current status of the activity, etc. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_ListActivityTypes.html

Request

data ListActivityTypes

Request constructor

listActivityTypes

Arguments

:: Text
latDomain

-> RegistrationStatus
latRegistrationStatus

-> ListActivityTypes

ListActivityTypes constructor.

The fields accessible through corresponding lenses are:

latDomain :: Text

latMaximumPageSize :: Maybe Natural

latName :: Maybe Text

latNextPageToken :: Maybe Text

latRegistrationStatus :: RegistrationStatus

latReverseOrder :: Maybe Bool

Request lenses

latDomain :: Lens' ListActivityTypes Text

The name of the domain in which the activity types have been registered.

latMaximumPageSize :: Lens' ListActivityTypes (Maybe Natural)

The maximum number of results that will be returned per call. nextPageToken can be used to obtain futher pages of results. The default is 100, which is the maximum allowed page size. You can, however, specify a page size smaller than 100.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

latName :: Lens' ListActivityTypes (Maybe Text)

If specified, only lists the activity types that have this name.

latNextPageToken :: Lens' ListActivityTypes (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

latRegistrationStatus :: Lens' ListActivityTypes RegistrationStatus

Specifies the registration status of the activity types to list.

latReverseOrder :: Lens' ListActivityTypes (Maybe Bool)

When set to true, returns the results in reverse order. By default, the results are returned in ascending alphabetical order by name of the activity types.

Response

data ListActivityTypesResponse

Response constructor

listActivityTypesResponse :: ListActivityTypesResponse

ListActivityTypesResponse constructor.

The fields accessible through corresponding lenses are:

latrNextPageToken :: Maybe Text

latrTypeInfos :: [ActivityTypeInfo]

Response lenses

latrNextPageToken :: Lens' ListActivityTypesResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

latrTypeInfos :: Lens' ListActivityTypesResponse [ActivityTypeInfo]

List of activity type information.