Network.AWS.SWF.ListClosedWorkflowExecutions

Description

Returns a list of closed workflow executions in the specified domain that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagFilter.tag : String constraint. The key is swf:tagFilter.tag .

typeFilter.name : String constraint. The key is swf:typeFilter.name .

typeFilter.version : String constraint. The key is swf:typeFilter.version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

This operation returns paginated results.

Creating a Request

listClosedWorkflowExecutions

Arguments

:: Text
lcweDomain

-> ListClosedWorkflowExecutions

Creates a value of ListClosedWorkflowExecutions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcweNextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

lcweExecutionFilter - If specified, only workflow executions matching the workflow ID specified in the filter are returned.

lcweCloseStatusFilter - If specified, only workflow executions that match this close status are listed. For example, if TERMINATED is specified, then only TERMINATED workflow executions are listed.

lcweTypeFilter - If specified, only executions of the type specified in the filter are returned.

lcweCloseTimeFilter - If specified, the workflow executions are included in the returned results based on whether their close times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their close times.

lcweReverseOrder - When set to true , returns the results in reverse order. By default the results are returned in descending order of the start or the close time of the executions.

lcweTagFilter - If specified, only executions that have the matching tag are listed.

lcweStartTimeFilter - If specified, the workflow executions are included in the returned results based on whether their start times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their start times.

lcweMaximumPageSize - The maximum number of results that are returned per call. nextPageToken can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size smaller than the maximum. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

lcweDomain - The name of the domain that contains the workflow executions to list.

data ListClosedWorkflowExecutions

See: listClosedWorkflowExecutions smart constructor.

Request Lenses

lcweNextPageToken :: Lens' ListClosedWorkflowExecutions (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

lcweExecutionFilter :: Lens' ListClosedWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the workflow ID specified in the filter are returned.

lcweCloseStatusFilter :: Lens' ListClosedWorkflowExecutions (Maybe CloseStatusFilter)

If specified, only workflow executions that match this close status are listed. For example, if TERMINATED is specified, then only TERMINATED workflow executions are listed.

lcweTypeFilter :: Lens' ListClosedWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, only executions of the type specified in the filter are returned.

lcweCloseTimeFilter :: Lens' ListClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, the workflow executions are included in the returned results based on whether their close times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their close times.

lcweReverseOrder :: Lens' ListClosedWorkflowExecutions (Maybe Bool)

When set to true , returns the results in reverse order. By default the results are returned in descending order of the start or the close time of the executions.

lcweTagFilter :: Lens' ListClosedWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have the matching tag are listed.

lcweStartTimeFilter :: Lens' ListClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, the workflow executions are included in the returned results based on whether their start times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their start times.

lcweMaximumPageSize :: Lens' ListClosedWorkflowExecutions (Maybe Natural)

The maximum number of results that are returned per call. nextPageToken can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size smaller than the maximum. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

lcweDomain :: Lens' ListClosedWorkflowExecutions Text

The name of the domain that contains the workflow executions to list.

Destructuring the Response

workflowExecutionInfos :: WorkflowExecutionInfos

Creates a value of WorkflowExecutionInfos with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

weiNextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

weiExecutionInfos - The list of workflow information structures.

data WorkflowExecutionInfos

Contains a paginated list of information about workflow executions.

See: workflowExecutionInfos smart constructor.

Response Lenses

weiNextPageToken :: Lens' WorkflowExecutionInfos (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

weiExecutionInfos :: Lens' WorkflowExecutionInfos [WorkflowExecutionInfo]

The list of workflow information structures.