Network.AWS.SWF.ListDomains

Description

Returns the list of domains registered in the account. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. The element must be set to arn:aws:swf::AccountID:domain/* , where AccountID is the account ID, with no dashes.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

This operation returns paginated results.

Creating a Request

listDomains

Arguments

:: RegistrationStatus
ldRegistrationStatus

-> ListDomains

Creates a value of ListDomains with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldNextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

ldReverseOrder - When set to true , returns the results in reverse order. By default, the results are returned in ascending alphabetical order by name of the domains.

ldMaximumPageSize - The maximum number of results that are returned per call. nextPageToken can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size smaller than the maximum. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

ldRegistrationStatus - Specifies the registration status of the domains to list.

data ListDomains

See: listDomains smart constructor.

Request Lenses

ldNextPageToken :: Lens' ListDomains (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

ldReverseOrder :: Lens' ListDomains (Maybe Bool)

When set to true , returns the results in reverse order. By default, the results are returned in ascending alphabetical order by name of the domains.

ldMaximumPageSize :: Lens' ListDomains (Maybe Natural)

The maximum number of results that are returned per call. nextPageToken can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size smaller than the maximum. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

ldRegistrationStatus :: Lens' ListDomains RegistrationStatus

Specifies the registration status of the domains to list.

Destructuring the Response

listDomainsResponse

Arguments

:: Int
ldrsResponseStatus

-> ListDomainsResponse

Creates a value of ListDomainsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldrsNextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

ldrsResponseStatus - -- | The response status code.

ldrsDomainInfos - A list of DomainInfo structures.

data ListDomainsResponse

Contains a paginated collection of DomainInfo structures.

See: listDomainsResponse smart constructor.

Response Lenses

ldrsNextPageToken :: Lens' ListDomainsResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

ldrsResponseStatus :: Lens' ListDomainsResponse Int

- | The response status code.

ldrsDomainInfos :: Lens' ListDomainsResponse [DomainInfo]

A list of DomainInfo structures.