Network.AWS.SWF.ListOpenWorkflowExecutions

Description

Returns a list of open workflow executions in the specified domain that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagFilter.tag : String constraint. The key is swf:tagFilter.tag .

typeFilter.name : String constraint. The key is swf:typeFilter.name .

typeFilter.version : String constraint. The key is swf:typeFilter.version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

This operation returns paginated results.

Creating a Request

listOpenWorkflowExecutions

Arguments

:: Text
loweDomain

-> ExecutionTimeFilter
loweStartTimeFilter

-> ListOpenWorkflowExecutions

Creates a value of ListOpenWorkflowExecutions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loweNextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

loweExecutionFilter - If specified, only workflow executions matching the workflow ID specified in the filter are returned.

loweTypeFilter - If specified, only executions of the type specified in the filter are returned.

loweReverseOrder - When set to true , returns the results in reverse order. By default the results are returned in descending order of the start time of the executions.

loweTagFilter - If specified, only executions that have the matching tag are listed.

loweMaximumPageSize - The maximum number of results that are returned per call. nextPageToken can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size smaller than the maximum. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

loweDomain - The name of the domain that contains the workflow executions to list.

loweStartTimeFilter - Workflow executions are included in the returned results based on whether their start times are within the range specified by this filter.

data ListOpenWorkflowExecutions

See: listOpenWorkflowExecutions smart constructor.

Request Lenses

loweNextPageToken :: Lens' ListOpenWorkflowExecutions (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

loweExecutionFilter :: Lens' ListOpenWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the workflow ID specified in the filter are returned.

loweTypeFilter :: Lens' ListOpenWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, only executions of the type specified in the filter are returned.

loweReverseOrder :: Lens' ListOpenWorkflowExecutions (Maybe Bool)

When set to true , returns the results in reverse order. By default the results are returned in descending order of the start time of the executions.

loweTagFilter :: Lens' ListOpenWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have the matching tag are listed.

loweMaximumPageSize :: Lens' ListOpenWorkflowExecutions (Maybe Natural)

The maximum number of results that are returned per call. nextPageToken can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size smaller than the maximum. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

loweDomain :: Lens' ListOpenWorkflowExecutions Text

The name of the domain that contains the workflow executions to list.

loweStartTimeFilter :: Lens' ListOpenWorkflowExecutions ExecutionTimeFilter

Workflow executions are included in the returned results based on whether their start times are within the range specified by this filter.

Destructuring the Response

workflowExecutionInfos :: WorkflowExecutionInfos

Creates a value of WorkflowExecutionInfos with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

weiNextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

weiExecutionInfos - The list of workflow information structures.

data WorkflowExecutionInfos

Contains a paginated list of information about workflow executions.

See: workflowExecutionInfos smart constructor.

Response Lenses

weiNextPageToken :: Lens' WorkflowExecutionInfos (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken . Keep all other arguments unchanged. The configured maximumPageSize determines how many results can be returned in a single call.

weiExecutionInfos :: Lens' WorkflowExecutionInfos [WorkflowExecutionInfo]

The list of workflow information structures.