Network.AWS.SWF.ListOpenWorkflowExecutions

Description

Returns a list of open workflow executions in the specified domain that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. tagFilter.tag: String constraint. The key is swf:tagFilter.tag. typeFilter.name: String constraint. The key is swf:typeFilter.name. typeFilter.version: String constraint. The key is swf:typeFilter.version. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see <http://docs.aws.amazon.com/amazonswf/latest/developerguide/swf-dev-iam.html Using IAM to Manage Access to Amazon SWF Workflows>.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_ListOpenWorkflowExecutions.html

Request

data ListOpenWorkflowExecutions

Request constructor

listOpenWorkflowExecutions

Arguments

:: Text
loweDomain

-> ExecutionTimeFilter
loweStartTimeFilter

-> ListOpenWorkflowExecutions

ListOpenWorkflowExecutions constructor.

The fields accessible through corresponding lenses are:

loweDomain :: Text

loweExecutionFilter :: Maybe WorkflowExecutionFilter

loweMaximumPageSize :: Maybe Natural

loweNextPageToken :: Maybe Text

loweReverseOrder :: Maybe Bool

loweStartTimeFilter :: ExecutionTimeFilter

loweTagFilter :: Maybe TagFilter

loweTypeFilter :: Maybe WorkflowTypeFilter

Request lenses

loweDomain :: Lens' ListOpenWorkflowExecutions Text

The name of the domain that contains the workflow executions to list.

loweExecutionFilter :: Lens' ListOpenWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the workflow id specified in the filter are returned.

loweMaximumPageSize :: Lens' ListOpenWorkflowExecutions (Maybe Natural)

The maximum number of results returned in each page. The default is 100, but the caller can override this value to a page size smaller than the default. You cannot specify a page size greater than 100. Note that the number of executions may be less than the maxiumum page size, in which case, the returned page will have fewer results than the maximumPageSize specified.

loweNextPageToken :: Lens' ListOpenWorkflowExecutions (Maybe Text)

If on a previous call to this method a NextPageToken was returned, the results are being paginated. To get the next page of results, repeat the call with the returned token and all other arguments unchanged.

loweReverseOrder :: Lens' ListOpenWorkflowExecutions (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in descending order of the start time of the executions.

loweStartTimeFilter :: Lens' ListOpenWorkflowExecutions ExecutionTimeFilter

Workflow executions are included in the returned results based on whether their start times are within the range specified by this filter.

loweTagFilter :: Lens' ListOpenWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have the matching tag are listed.

loweTypeFilter :: Lens' ListOpenWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, only executions of the type specified in the filter are returned.

Response

data ListOpenWorkflowExecutionsResponse

Response constructor

listOpenWorkflowExecutionsResponse :: ListOpenWorkflowExecutionsResponse

ListOpenWorkflowExecutionsResponse constructor.

The fields accessible through corresponding lenses are:

lowerExecutionInfos :: [WorkflowExecutionInfo]

lowerNextPageToken :: Maybe Text

Response lenses

lowerExecutionInfos :: Lens' ListOpenWorkflowExecutionsResponse [WorkflowExecutionInfo]

The list of workflow information structures.

lowerNextPageToken :: Lens' ListOpenWorkflowExecutionsResponse (Maybe Text)

The token of the next page in the result. If set, the results have more than one page. The next page can be retrieved by repeating the request with this token and all other arguments unchanged.