Network.AWS.SWF.ListOpenWorkflowExecutions

Description

Returns a list of open workflow executions in the specified domain that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the following parameters by using a Condition element with the appropriate keys. 'tagFilter.tag': String constraint. The key is 'swf:tagFilter.tag'. 'typeFilter.name': String constraint. The key is 'swf:typeFilter.name'. 'typeFilter.version': String constraint. The key is 'swf:typeFilter.version'. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to AmazonSWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_ListOpenWorkflowExecutions.html

Request

data ListOpenWorkflowExecutions

Request constructor

listOpenWorkflowExecutions

Arguments

:: Text
loweDomain

-> ExecutionTimeFilter
loweStartTimeFilter

-> ListOpenWorkflowExecutions

ListOpenWorkflowExecutions constructor.

The fields accessible through corresponding lenses are:

loweDomain :: Text

loweExecutionFilter :: Maybe WorkflowExecutionFilter

loweMaximumPageSize :: Maybe Natural

loweNextPageToken :: Maybe Text

loweReverseOrder :: Maybe Bool

loweStartTimeFilter :: ExecutionTimeFilter

loweTagFilter :: Maybe TagFilter

loweTypeFilter :: Maybe WorkflowTypeFilter

Request lenses

loweDomain :: Lens' ListOpenWorkflowExecutions Text

The name of the domain that contains the workflow executions to list.

loweExecutionFilter :: Lens' ListOpenWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the workflow id specified in the filter are returned.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

loweMaximumPageSize :: Lens' ListOpenWorkflowExecutions (Maybe Natural)

The maximum number of results that will be returned per call. nextPageToken can be used to obtain futher pages of results. The default is 100, which is the maximum allowed page size. You can, however, specify a page size smaller than 100.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

loweNextPageToken :: Lens' ListOpenWorkflowExecutions (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

loweReverseOrder :: Lens' ListOpenWorkflowExecutions (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in descending order of the start time of the executions.

loweStartTimeFilter :: Lens' ListOpenWorkflowExecutions ExecutionTimeFilter

Workflow executions are included in the returned results based on whether their start times are within the range specified by this filter.

loweTagFilter :: Lens' ListOpenWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have the matching tag are listed.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

loweTypeFilter :: Lens' ListOpenWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, only executions of the type specified in the filter are returned.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

Response

data ListOpenWorkflowExecutionsResponse

Response constructor

listOpenWorkflowExecutionsResponse :: ListOpenWorkflowExecutionsResponse

ListOpenWorkflowExecutionsResponse constructor.

The fields accessible through corresponding lenses are:

lowerExecutionInfos :: [WorkflowExecutionInfo]

lowerNextPageToken :: Maybe Text

Response lenses

lowerExecutionInfos :: Lens' ListOpenWorkflowExecutionsResponse [WorkflowExecutionInfo]

The list of workflow information structures.

lowerNextPageToken :: Lens' ListOpenWorkflowExecutionsResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.