Network.AWS.SWF.ListWorkflowTypes

Description

Returns information about workflow types in the specified domain. The results may be split into multiple pages that can be retrieved by making the call repeatedly. Access Control You can use IAM policies to control this action's access to Amazon SWF resources as follows: Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails by throwing OperationNotPermitted. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows. ListWorkflowTypes Example POST HTTP1.1 Host: swf.us-east-1.amazonaws.com User-Agent: Mozilla5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.2.25) Gecko20111212 Firefox3.6.25 ( .NET CLR 3.5.30729; .NET4.0E) Accept: applicationjson, textjavascript, ** Accept-Language: en-us,en;q=0.5 Accept-Encoding: gzip,deflate Accept-Charset: ISO-8859-1,utf-8;q=0.7,*;q=0.7 Keep-Alive: 115 Connection: keep-alive Content-Type: application/x-amz-json-1.0 X-Requested-With: XMLHttpRequest X-Amz-Date: Sun, 15 Jan 2012 22:25:43 GMT X-Amz-Target: SimpleWorkflowService.ListWorkflowTypes Content-Encoding: amz-1.0 X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE,Algorithm=HmacSHA256,SignedHeaders=Host;X-Amz-Date;X-Amz-Target;Content-Encoding,Signature=uleWQSyVVf0+aG50IoBJG5h0hzxNFNT97Mkn/FSCQ+Q= Referer: http:/swf.us-east-1.amazonaws.comexplorer/index.html Content-Length: 110 Pragma: no-cache Cache-Control: no-cache {"domain": "867530901", "registrationStatus": REGISTERED, "maximumPageSize": 50, "reverseOrder": true} HTTP/1.1 200 OK Content-Length: 174 Content-Type: application/json x-amzn-RequestId: dcde6719-3fc7-11e1-9e8f-57bb03e21482 {"typeInfos": [ {"creationDate": 1326481174.027, "description": "Handle customer orders", "status": REGISTERED, "workflowType": {"name": "customerOrderWorkflow", "version": "1.0"} } ] }.

Request

data ListWorkflowTypes

Request constructor

mkListWorkflowTypes

Arguments

:: Text
lwtDomain

-> RegistrationStatus
lwtRegistrationStatus

-> ListWorkflowTypes

Smart constructor for the minimum required parameters to construct a valid ListWorkflowTypes request.

The fields accessible through corresponding lenses are:

Domain :: Text

Name :: Maybe Text

RegistrationStatus :: RegistrationStatus

NextPageToken :: Maybe Text

MaximumPageSize :: Maybe Integer

ReverseOrder :: Maybe Bool

Request lenses

lwtDomain :: Lens' ListWorkflowTypes Text

The name of the domain in which the workflow types have been registered.

lwtName :: Lens' ListWorkflowTypes (Maybe Text)

If specified, lists the workflow type with this name.

lwtRegistrationStatus :: Lens' ListWorkflowTypes RegistrationStatus

Specifies the registration status of the workflow types to list.

lwtNextPageToken :: Lens' ListWorkflowTypes (Maybe Text)

If on a previous call to this method a NextPageToken was returned, the results are being paginated. To get the next page of results, repeat the call with the returned token and all other arguments unchanged.

lwtMaximumPageSize :: Lens' ListWorkflowTypes (Maybe Integer)

The maximum number of results returned in each page. The default is 100, but the caller can override this value to a page size smaller than the default. You cannot specify a page size greater than 100. Note that the number of types may be less than the maxiumum page size, in which case, the returned page will have fewer results than the maximumPageSize specified.

lwtReverseOrder :: Lens' ListWorkflowTypes (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in ascending alphabetical order of the name of the workflow types.

Response

data ListWorkflowTypesResponse

Contains a paginated list of information structures about workflow types.

Response constructor

mkListWorkflowTypesResponse

Arguments

:: [WorkflowTypeInfo]
lwtrTypeInfos

-> ListWorkflowTypesResponse

Smart constructor for the minimum required parameters to construct a valid ListWorkflowTypesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TypeInfos :: [WorkflowTypeInfo]

NextPageToken :: Maybe Text

Response lenses

lwtrTypeInfos :: Lens' ListWorkflowTypesResponse [WorkflowTypeInfo]

The list of workflow type information.

lwtrNextPageToken :: Lens' ListWorkflowTypesResponse (Maybe Text)

The token for the next page of type information. If set then the list consists of more than one page. You can retrieve the next page by repeating the request (that returned the structure) with the this token and all other arguments unchanged.