Network.AWS.SWF.ListWorkflowTypes

Description

Returns information about workflow types in the specified domain. The results may be split into multiple pages that can be retrieved by making the call repeatedly.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_ListWorkflowTypes.html

Request

data ListWorkflowTypes

Request constructor

listWorkflowTypes

Arguments

:: Text
lwtDomain

-> RegistrationStatus
lwtRegistrationStatus

-> ListWorkflowTypes

ListWorkflowTypes constructor.

The fields accessible through corresponding lenses are:

lwtDomain :: Text

lwtMaximumPageSize :: Maybe Natural

lwtName :: Maybe Text

lwtNextPageToken :: Maybe Text

lwtRegistrationStatus :: RegistrationStatus

lwtReverseOrder :: Maybe Bool

Request lenses

lwtDomain :: Lens' ListWorkflowTypes Text

The name of the domain in which the workflow types have been registered.

lwtMaximumPageSize :: Lens' ListWorkflowTypes (Maybe Natural)

The maximum number of results that will be returned per call. nextPageToken can be used to obtain futher pages of results. The default is 100, which is the maximum allowed page size. You can, however, specify a page size smaller than 100.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

lwtName :: Lens' ListWorkflowTypes (Maybe Text)

If specified, lists the workflow type with this name.

lwtNextPageToken :: Lens' ListWorkflowTypes (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

lwtRegistrationStatus :: Lens' ListWorkflowTypes RegistrationStatus

Specifies the registration status of the workflow types to list.

lwtReverseOrder :: Lens' ListWorkflowTypes (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in ascending alphabetical order of the name of the workflow types.

Response

data ListWorkflowTypesResponse

Response constructor

listWorkflowTypesResponse :: ListWorkflowTypesResponse

ListWorkflowTypesResponse constructor.

The fields accessible through corresponding lenses are:

lwtrNextPageToken :: Maybe Text

lwtrTypeInfos :: [WorkflowTypeInfo]

Response lenses

lwtrNextPageToken :: Lens' ListWorkflowTypesResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

lwtrTypeInfos :: Lens' ListWorkflowTypesResponse [WorkflowTypeInfo]

The list of workflow type information.