Network.AWS.SWF.PollForActivityTask

Description

Used by workers to get an ActivityTask from the specified activity taskList . This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available. The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll returns an empty result. An empty result, in this context, means that an ActivityTask is returned, but that the value of taskToken is an empty string. If a task is returned, the worker should use its type to identify and process it correctly.

Important: Workers should set their client side socket timeout to at least 70 seconds (10 seconds higher than the maximum time service may hold the poll request).

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the taskList.name parameter by using a Condition element with the swf:taskList.name key to allow the action to access only certain task lists.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

pollForActivityTask

Arguments

:: Text
pfatDomain

-> TaskList
pfatTaskList

-> PollForActivityTask

Creates a value of PollForActivityTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pfatIdentity - Identity of the worker making the request, recorded in the ActivityTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

pfatDomain - The name of the domain that contains the task lists being polled.

pfatTaskList - Specifies the task list to poll for activity tasks. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

data PollForActivityTask

See: pollForActivityTask smart constructor.

Request Lenses

pfatIdentity :: Lens' PollForActivityTask (Maybe Text)

Identity of the worker making the request, recorded in the ActivityTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

pfatDomain :: Lens' PollForActivityTask Text

The name of the domain that contains the task lists being polled.

pfatTaskList :: Lens' PollForActivityTask TaskList

Specifies the task list to poll for activity tasks. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

Destructuring the Response

pollForActivityTaskResponse

Arguments

:: Int
pfatrsResponseStatus

-> Integer
pfatrsStartedEventId

-> PollForActivityTaskResponse

Creates a value of PollForActivityTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pfatrsActivityType - The type of this activity task.

pfatrsActivityId - The unique ID of the task.

pfatrsInput - The inputs provided when the activity task was scheduled. The form of the input is user defined and should be meaningful to the activity implementation.

pfatrsTaskToken - The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

pfatrsWorkflowExecution - The workflow execution that started this activity task.

pfatrsResponseStatus - -- | The response status code.

pfatrsStartedEventId - The ID of the ActivityTaskStarted event recorded in the history.

data PollForActivityTaskResponse

Unit of work sent to an activity worker.

See: pollForActivityTaskResponse smart constructor.

Response Lenses

pfatrsActivityType :: Lens' PollForActivityTaskResponse (Maybe ActivityType)

The type of this activity task.

pfatrsActivityId :: Lens' PollForActivityTaskResponse (Maybe Text)

The unique ID of the task.

pfatrsInput :: Lens' PollForActivityTaskResponse (Maybe Text)

The inputs provided when the activity task was scheduled. The form of the input is user defined and should be meaningful to the activity implementation.

pfatrsTaskToken :: Lens' PollForActivityTaskResponse (Maybe Text)

The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

pfatrsWorkflowExecution :: Lens' PollForActivityTaskResponse (Maybe WorkflowExecution)

The workflow execution that started this activity task.

pfatrsResponseStatus :: Lens' PollForActivityTaskResponse Int

- | The response status code.

pfatrsStartedEventId :: Lens' PollForActivityTaskResponse Integer

The ID of the ActivityTaskStarted event recorded in the history.