Network.AWS.SWF.PollForActivityTask

Description

Used by workers to get an ActivityTask from the specified activity taskList. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available. The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll will return an empty result. An empty result, in this context, means that an ActivityTask is returned, but that the value of taskToken is an empty string. If a task is returned, the worker should use its type to identify and process it correctly.

Workers should set their client side socket timeout to at least 70 seconds (10 seconds higher than the maximum time service may hold the poll request). Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the 'taskList.name' parameter by using a Condition element with the 'swf:taskList.name' key to allow the action to access only certain task lists. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_PollForActivityTask.html

Request

data PollForActivityTask

Request constructor

pollForActivityTask

Arguments

:: Text
pfatDomain

-> TaskList
pfatTaskList

-> PollForActivityTask

PollForActivityTask constructor.

The fields accessible through corresponding lenses are:

pfatDomain :: Text

pfatIdentity :: Maybe Text

pfatTaskList :: TaskList

Request lenses

pfatDomain :: Lens' PollForActivityTask Text

The name of the domain that contains the task lists being polled.

pfatIdentity :: Lens' PollForActivityTask (Maybe Text)

Identity of the worker making the request, recorded in the ActivityTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

pfatTaskList :: Lens' PollForActivityTask TaskList

Specifies the task list to poll for activity tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), '|' (vertical bar), or any control characters (u0000-u001f | u007f - u009f). Also, it must not contain the literal string quotarnquot.

Response

data PollForActivityTaskResponse

Response constructor

pollForActivityTaskResponse

Arguments

:: Text
pfatrTaskToken

-> Text
pfatrActivityId

-> Integer
pfatrStartedEventId

-> WorkflowExecution
pfatrWorkflowExecution

-> ActivityType
pfatrActivityType

-> PollForActivityTaskResponse

PollForActivityTaskResponse constructor.

The fields accessible through corresponding lenses are:

pfatrActivityId :: Text

pfatrActivityType :: ActivityType

pfatrInput :: Maybe Text

pfatrStartedEventId :: Integer

pfatrTaskToken :: Text

pfatrWorkflowExecution :: WorkflowExecution

Response lenses

pfatrActivityId :: Lens' PollForActivityTaskResponse Text

The unique ID of the task.

pfatrActivityType :: Lens' PollForActivityTaskResponse ActivityType

The type of this activity task.

pfatrInput :: Lens' PollForActivityTaskResponse (Maybe Text)

The inputs provided when the activity task was scheduled. The form of the input is user defined and should be meaningful to the activity implementation.

pfatrStartedEventId :: Lens' PollForActivityTaskResponse Integer

The id of the ActivityTaskStarted event recorded in the history.

pfatrTaskToken :: Lens' PollForActivityTaskResponse Text

The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

pfatrWorkflowExecution :: Lens' PollForActivityTaskResponse WorkflowExecution

The workflow execution that started this activity task.