Network.AWS.SWF.PollForDecisionTask

Description

Used by deciders to get a DecisionTask from the specified decision taskList. A decision task may be returned for any open workflow execution that is using the specified task list. The task includes a paginated view of the history of the workflow execution. The decider should use the workflow type and the history to determine how to properly handle the task.

This action initiates a long poll, where the service holds the HTTP connection open and responds as soon a task becomes available. If no decision task is available in the specified task list before the timeout of 60 seconds expires, an empty result is returned. An empty result, in this context, means that a DecisionTask is returned, but that the value of taskToken is an empty string.

Deciders should set their client side socket timeout to at least 70 seconds (10 seconds higher than the timeout). Because the number of workflow history events for a single workflow execution might be very large, the result returned might be split up across a number of pages. To retrieve subsequent pages, make additional calls to PollForDecisionTask using the nextPageToken returned by the initial call. Note that you do not call GetWorkflowExecutionHistory with this nextPageToken. Instead, call PollForDecisionTask again. Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. Constrain the 'taskList.name' parameter by using a Condition element with the 'swf:taskList.name' key to allow the action to access only certain task lists. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_PollForDecisionTask.html

Request

data PollForDecisionTask

Request constructor

pollForDecisionTask

Arguments

:: Text
pfdtDomain

-> TaskList
pfdtTaskList

-> PollForDecisionTask

PollForDecisionTask constructor.

The fields accessible through corresponding lenses are:

pfdtDomain :: Text

pfdtIdentity :: Maybe Text

pfdtMaximumPageSize :: Maybe Natural

pfdtNextPageToken :: Maybe Text

pfdtReverseOrder :: Maybe Bool

pfdtTaskList :: TaskList

Request lenses

pfdtDomain :: Lens' PollForDecisionTask Text

The name of the domain containing the task lists to poll.

pfdtIdentity :: Lens' PollForDecisionTask (Maybe Text)

Identity of the decider making the request, which is recorded in the DecisionTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

pfdtMaximumPageSize :: Lens' PollForDecisionTask (Maybe Natural)

The maximum number of results that will be returned per call. nextPageToken can be used to obtain futher pages of results. The default is 100, which is the maximum allowed page size. You can, however, specify a page size smaller than 100.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

pfdtNextPageToken :: Lens' PollForDecisionTask (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

The nextPageToken returned by this action cannot be used with GetWorkflowExecutionHistory to get the next page. You must call PollForDecisionTask again (with the nextPageToken) to retrieve the next page of history records. Calling PollForDecisionTask with a nextPageToken will not return a new decision task..

pfdtReverseOrder :: Lens' PollForDecisionTask (Maybe Bool)

When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimestamp of the events.

pfdtTaskList :: Lens' PollForDecisionTask TaskList

Specifies the task list to poll for decision tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), '|' (vertical bar), or any control characters (u0000-u001f | u007f - u009f). Also, it must not contain the literal string quotarnquot.

Response

data PollForDecisionTaskResponse

Response constructor

pollForDecisionTaskResponse

Arguments

:: Text
pfdtrTaskToken

-> Integer
pfdtrStartedEventId

-> WorkflowExecution
pfdtrWorkflowExecution

-> WorkflowType
pfdtrWorkflowType

-> PollForDecisionTaskResponse

PollForDecisionTaskResponse constructor.

The fields accessible through corresponding lenses are:

pfdtrEvents :: [HistoryEvent]

pfdtrNextPageToken :: Maybe Text

pfdtrPreviousStartedEventId :: Maybe Integer

pfdtrStartedEventId :: Integer

pfdtrTaskToken :: Text

pfdtrWorkflowExecution :: WorkflowExecution

pfdtrWorkflowType :: WorkflowType

Response lenses

pfdtrEvents :: Lens' PollForDecisionTaskResponse [HistoryEvent]

A paginated list of history events of the workflow execution. The decider uses this during the processing of the decision task.

pfdtrNextPageToken :: Lens' PollForDecisionTaskResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

pfdtrPreviousStartedEventId :: Lens' PollForDecisionTaskResponse (Maybe Integer)

The id of the DecisionTaskStarted event of the previous decision task of this workflow execution that was processed by the decider. This can be used to determine the events in the history new since the last decision task received by the decider.

pfdtrStartedEventId :: Lens' PollForDecisionTaskResponse Integer

The id of the DecisionTaskStarted event recorded in the history.

pfdtrTaskToken :: Lens' PollForDecisionTaskResponse Text

The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

pfdtrWorkflowExecution :: Lens' PollForDecisionTaskResponse WorkflowExecution

The workflow execution for which this decision task was created.

pfdtrWorkflowType :: Lens' PollForDecisionTaskResponse WorkflowType

The type of the workflow execution for which this decision task was created.