Network.AWS.SWF.RecordActivityTaskHeartbeat

Description

Used by activity workers to report to the service that the ActivityTask represented by the specified taskToken is still making progress. The worker can also specify details of the progress, for example percent complete, using the details parameter. This action can also be used by the worker as a mechanism to check if cancellation is being requested for the activity task. If a cancellation is being attempted for the specified task, then the boolean cancelRequested flag returned by the service is set to true .

This action resets the taskHeartbeatTimeout clock. The taskHeartbeatTimeout is specified in RegisterActivityType .

This action doesn't in itself create an event in the workflow execution history. However, if the task times out, the workflow execution history contains a ActivityTaskTimedOut event that contains the information from the last heartbeat generated by the activity worker.

Important: If the cancelRequested flag returns true , a cancellation is being attempted. If the worker can cancel the activity, it should respond with RespondActivityTaskCanceled . Otherwise, it should ignore the cancellation request.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

recordActivityTaskHeartbeat

Arguments

:: Text
rathTaskToken

-> RecordActivityTaskHeartbeat

Creates a value of RecordActivityTaskHeartbeat with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rathDetails - If specified, contains details about the progress of the task.

rathTaskToken - The taskToken of the ActivityTask . Important: taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

data RecordActivityTaskHeartbeat

See: recordActivityTaskHeartbeat smart constructor.

Request Lenses

rathDetails :: Lens' RecordActivityTaskHeartbeat (Maybe Text)

If specified, contains details about the progress of the task.

rathTaskToken :: Lens' RecordActivityTaskHeartbeat Text

The taskToken of the ActivityTask . Important: taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

Destructuring the Response

recordActivityTaskHeartbeatResponse

Arguments

:: Int
rathrsResponseStatus

-> Bool
rathrsCancelRequested

-> RecordActivityTaskHeartbeatResponse

Creates a value of RecordActivityTaskHeartbeatResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rathrsResponseStatus - -- | The response status code.

rathrsCancelRequested - Set to true if cancellation of the task is requested.

data RecordActivityTaskHeartbeatResponse

Status information about an activity task.

See: recordActivityTaskHeartbeatResponse smart constructor.

Response Lenses

rathrsResponseStatus :: Lens' RecordActivityTaskHeartbeatResponse Int

- | The response status code.

rathrsCancelRequested :: Lens' RecordActivityTaskHeartbeatResponse Bool

Set to true if cancellation of the task is requested.