Network.AWS.SWF.RegisterActivityType

Description

Registers a new activity type along with its configuration settings in the specified domain.

Important: A TypeAlreadyExists fault is returned if the type already exists in the domain. You cannot change any configuration settings of the type after its registration, and it must be registered as a new version.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

defaultTaskList.name : String constraint. The key is swf:defaultTaskList.name .

name : String constraint. The key is swf:name .

version : String constraint. The key is swf:version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

registerActivityType

Arguments

:: Text
ratDomain

-> Text
ratName

-> Text
ratVersion

-> RegisterActivityType

Creates a value of RegisterActivityType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ratDefaultTaskScheduleToStartTimeout - If set, specifies the default maximum duration that a task of this activity type can wait before being assigned to a worker. This default can be overridden when scheduling an activity task using the ScheduleActivityTask Decision . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

ratDefaultTaskList - If set, specifies the default task list to use for scheduling tasks of this activity type. This default task list is used if a task list isn't provided when a task is scheduled through the ScheduleActivityTask Decision .

ratDefaultTaskPriority - The default task priority to assign to the activity type. If not assigned, then 0 is used. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority. For more information about setting task priority, see Setting Task Priority in the in the /Amazon SWF Developer Guide/ . .

ratDefaultTaskHeartbeatTimeout - If set, specifies the default maximum time before which a worker processing a task of this type must report progress by calling RecordActivityTaskHeartbeat . If the timeout is exceeded, the activity task is automatically timed out. This default can be overridden when scheduling an activity task using the ScheduleActivityTask Decision . If the activity worker subsequently attempts to record a heartbeat or returns a result, the activity worker receives an UnknownResource fault. In this case, Amazon SWF no longer considers the activity task to be valid; the activity worker should clean up the activity task. The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

ratDefaultTaskScheduleToCloseTimeout - If set, specifies the default maximum duration for a task of this activity type. This default can be overridden when scheduling an activity task using the ScheduleActivityTask Decision . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

ratDefaultTaskStartToCloseTimeout - If set, specifies the default maximum duration that a worker can take to process tasks of this activity type. This default can be overridden when scheduling an activity task using the ScheduleActivityTask Decision . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

ratDescription - A textual description of the activity type.

ratDomain - The name of the domain in which this activity is to be registered.

ratName - The name of the activity type within the domain. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

ratVersion - The version of the activity type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

data RegisterActivityType

See: registerActivityType smart constructor.

Request Lenses

ratDefaultTaskScheduleToStartTimeout :: Lens' RegisterActivityType (Maybe Text)

If set, specifies the default maximum duration that a task of this activity type can wait before being assigned to a worker. This default can be overridden when scheduling an activity task using the ScheduleActivityTask Decision . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

ratDefaultTaskList :: Lens' RegisterActivityType (Maybe TaskList)

If set, specifies the default task list to use for scheduling tasks of this activity type. This default task list is used if a task list isn't provided when a task is scheduled through the ScheduleActivityTask Decision .

ratDefaultTaskPriority :: Lens' RegisterActivityType (Maybe Text)

The default task priority to assign to the activity type. If not assigned, then 0 is used. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority. For more information about setting task priority, see Setting Task Priority in the in the /Amazon SWF Developer Guide/ . .

ratDefaultTaskHeartbeatTimeout :: Lens' RegisterActivityType (Maybe Text)

If set, specifies the default maximum time before which a worker processing a task of this type must report progress by calling RecordActivityTaskHeartbeat . If the timeout is exceeded, the activity task is automatically timed out. This default can be overridden when scheduling an activity task using the ScheduleActivityTask Decision . If the activity worker subsequently attempts to record a heartbeat or returns a result, the activity worker receives an UnknownResource fault. In this case, Amazon SWF no longer considers the activity task to be valid; the activity worker should clean up the activity task. The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

ratDefaultTaskScheduleToCloseTimeout :: Lens' RegisterActivityType (Maybe Text)

If set, specifies the default maximum duration for a task of this activity type. This default can be overridden when scheduling an activity task using the ScheduleActivityTask Decision . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

ratDefaultTaskStartToCloseTimeout :: Lens' RegisterActivityType (Maybe Text)

If set, specifies the default maximum duration that a worker can take to process tasks of this activity type. This default can be overridden when scheduling an activity task using the ScheduleActivityTask Decision . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

ratDescription :: Lens' RegisterActivityType (Maybe Text)

A textual description of the activity type.

ratDomain :: Lens' RegisterActivityType Text

The name of the domain in which this activity is to be registered.

ratName :: Lens' RegisterActivityType Text

The name of the activity type within the domain. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

ratVersion :: Lens' RegisterActivityType Text

The version of the activity type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

Destructuring the Response

registerActivityTypeResponse :: RegisterActivityTypeResponse

Creates a value of RegisterActivityTypeResponse with the minimum fields required to make a request.

data RegisterActivityTypeResponse

See: registerActivityTypeResponse smart constructor.