Network.AWS.SWF.RegisterDomain

Description

Registers a new domain.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

You cannot use an IAM policy to control domain access for this action. The name of the domain being registered is available as the resource of this action. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_RegisterDomain.html

Request

data RegisterDomain

Request constructor

registerDomain

Arguments

:: Text
rdName

-> Text
rdWorkflowExecutionRetentionPeriodInDays

-> RegisterDomain

RegisterDomain constructor.

The fields accessible through corresponding lenses are:

rdDescription :: Maybe Text

rdName :: Text

rdWorkflowExecutionRetentionPeriodInDays :: Text

Request lenses

rdDescription :: Lens' RegisterDomain (Maybe Text)

A text description of the domain.

rdName :: Lens' RegisterDomain Text

Name of the domain to register. The name must be unique in the region that the domain is registered in.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), '|' (vertical bar), or any control characters (u0000-u001f | u007f - u009f). Also, it must not contain the literal string quotarnquot.

rdWorkflowExecutionRetentionPeriodInDays :: Lens' RegisterDomain Text

The duration (in days) that records and histories of workflow executions on the domain should be kept by the service. After the retention period, the workflow execution is not available in the results of visibility calls.

If you pass the value NONE or '0' (zero), then the workflow execution history will not be retained. As soon as the workflow execution completes, the execution record and its history are deleted.

The maximum workflow execution retention period is 90 days. For more information about Amazon SWF service limits, see: Amazon SWF Service Limits in the Amazon SWF Developer Guide.

Response

data RegisterDomainResponse

Response constructor

registerDomainResponse :: RegisterDomainResponse

RegisterDomainResponse constructor.