Network.AWS.SWF.RegisterWorkflowType

Description

Registers a new workflow type and its configuration settings in the specified domain.

The retention period for the workflow history is set by the RegisterDomain action.

Important: If the type already exists, then a TypeAlreadyExists fault is returned. You cannot change the configuration settings of a workflow type once it is registered and it must be registered as a new version.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

defaultTaskList.name : String constraint. The key is swf:defaultTaskList.name .

name : String constraint. The key is swf:name .

version : String constraint. The key is swf:version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

registerWorkflowType

Arguments

:: Text
rwtDomain

-> Text
rwtName

-> Text
rwtVersion

-> RegisterWorkflowType

Creates a value of RegisterWorkflowType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rwtDefaultLambdaRole - The default IAM role attached to this workflow type.

rwtDefaultChildPolicy - If set, specifies the default policy to use for the child workflow executions when a workflow execution of this type is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision . The supported child policies are: * TERMINATE – The child executions are terminated. * REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. * ABANDON – No action is taken. The child executions continue to run.

rwtDefaultTaskList - If set, specifies the default task list to use for scheduling decision tasks for executions of this workflow type. This default is used only if a task list isn't provided when starting the execution through the StartWorkflowExecution Action or StartChildWorkflowExecution Decision .

rwtDefaultTaskPriority - The default task priority to assign to the workflow type. If not assigned, then 0 is used. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority. For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide .

rwtDefaultExecutionStartToCloseTimeout - If set, specifies the default maximum duration for executions of this workflow type. You can override this default when starting an execution through the StartWorkflowExecution Action or StartChildWorkflowExecution Decision . The duration is specified in seconds; an integer greater than or equal to 0. Unlike some of the other timeout parameters in Amazon SWF, you cannot specify a value of NONE for defaultExecutionStartToCloseTimeout ; there is a one-year max limit on the time that a workflow execution can run. Exceeding this limit always causes the workflow execution to time out.

rwtDefaultTaskStartToCloseTimeout - If set, specifies the default maximum duration of decision tasks for this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

rwtDescription - Textual description of the workflow type.

rwtDomain - The name of the domain in which to register the workflow type.

rwtName - The name of the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

rwtVersion - The version of the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

data RegisterWorkflowType

See: registerWorkflowType smart constructor.

Request Lenses

rwtDefaultLambdaRole :: Lens' RegisterWorkflowType (Maybe Text)

The default IAM role attached to this workflow type.

rwtDefaultChildPolicy :: Lens' RegisterWorkflowType (Maybe ChildPolicy)

If set, specifies the default policy to use for the child workflow executions when a workflow execution of this type is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision . The supported child policies are: * TERMINATE – The child executions are terminated. * REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. * ABANDON – No action is taken. The child executions continue to run.

rwtDefaultTaskList :: Lens' RegisterWorkflowType (Maybe TaskList)

If set, specifies the default task list to use for scheduling decision tasks for executions of this workflow type. This default is used only if a task list isn't provided when starting the execution through the StartWorkflowExecution Action or StartChildWorkflowExecution Decision .

rwtDefaultTaskPriority :: Lens' RegisterWorkflowType (Maybe Text)

The default task priority to assign to the workflow type. If not assigned, then 0 is used. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority. For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide .

rwtDefaultExecutionStartToCloseTimeout :: Lens' RegisterWorkflowType (Maybe Text)

If set, specifies the default maximum duration for executions of this workflow type. You can override this default when starting an execution through the StartWorkflowExecution Action or StartChildWorkflowExecution Decision . The duration is specified in seconds; an integer greater than or equal to 0. Unlike some of the other timeout parameters in Amazon SWF, you cannot specify a value of NONE for defaultExecutionStartToCloseTimeout ; there is a one-year max limit on the time that a workflow execution can run. Exceeding this limit always causes the workflow execution to time out.

rwtDefaultTaskStartToCloseTimeout :: Lens' RegisterWorkflowType (Maybe Text)

If set, specifies the default maximum duration of decision tasks for this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

rwtDescription :: Lens' RegisterWorkflowType (Maybe Text)

Textual description of the workflow type.

rwtDomain :: Lens' RegisterWorkflowType Text

The name of the domain in which to register the workflow type.

rwtName :: Lens' RegisterWorkflowType Text

The name of the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

rwtVersion :: Lens' RegisterWorkflowType Text

The version of the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

Destructuring the Response

registerWorkflowTypeResponse :: RegisterWorkflowTypeResponse

Creates a value of RegisterWorkflowTypeResponse with the minimum fields required to make a request.

data RegisterWorkflowTypeResponse

See: registerWorkflowTypeResponse smart constructor.