Network.AWS.SWF.RequestCancelWorkflowExecution

Description

Records a WorkflowExecutionCancelRequested event in the currently running workflow execution identified by the given domain, workflowId, and runId. This logically requests the cancellation of the workflow execution as a whole. It is up to the decider to take appropriate actions when it receives an execution history with this event.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

requestCancelWorkflowExecution

Arguments

:: Text
rcweDomain

-> Text
rcweWorkflowId

-> RequestCancelWorkflowExecution

Creates a value of RequestCancelWorkflowExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcweRunId - The runId of the workflow execution to cancel.

rcweDomain - The name of the domain containing the workflow execution to cancel.

rcweWorkflowId - The workflowId of the workflow execution to cancel.

data RequestCancelWorkflowExecution

See: requestCancelWorkflowExecution smart constructor.

Request Lenses

rcweRunId :: Lens' RequestCancelWorkflowExecution (Maybe Text)

The runId of the workflow execution to cancel.

rcweDomain :: Lens' RequestCancelWorkflowExecution Text

The name of the domain containing the workflow execution to cancel.

rcweWorkflowId :: Lens' RequestCancelWorkflowExecution Text

The workflowId of the workflow execution to cancel.

Destructuring the Response

requestCancelWorkflowExecutionResponse :: RequestCancelWorkflowExecutionResponse

Creates a value of RequestCancelWorkflowExecutionResponse with the minimum fields required to make a request.

data RequestCancelWorkflowExecutionResponse

See: requestCancelWorkflowExecutionResponse smart constructor.