Network.AWS.SWF.RespondActivityTaskCompleted

Description

Used by workers to tell the service that the ActivityTask identified by the taskToken completed successfully with a result (if provided). The result appears in the ActivityTaskCompleted event in the workflow history.

If the requested task does not complete successfully, use RespondActivityTaskFailed instead. If the worker finds that the task is canceled through the canceled flag returned by RecordActivityTaskHeartbeat, it should cancel the task, clean up and then call RespondActivityTaskCanceled. A task is considered open from the time that it is scheduled until it is closed. Therefore a task is reported as open while a worker is processing it. A task is closed after it has been specified in a call to RespondActivityTaskCompleted, RespondActivityTaskCanceled, RespondActivityTaskFailed, or the task has timed out.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. Use an Action element to allow or deny permission to call this action. You cannot use an IAM policy to constrain this action's parameters. If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAMto Manage Access to Amazon SWF Workflows.

http://docs.aws.amazon.com/amazonswf/latest/apireference/API_RespondActivityTaskCompleted.html

Request

data RespondActivityTaskCompleted

Request constructor

respondActivityTaskCompleted

Arguments

:: Text
ratcTaskToken

-> RespondActivityTaskCompleted

RespondActivityTaskCompleted constructor.

The fields accessible through corresponding lenses are:

ratcResult :: Maybe Text

ratcTaskToken :: Text

Request lenses

ratcResult :: Lens' RespondActivityTaskCompleted (Maybe Text)

The result of the activity task. It is a free form string that is implementation specific.

ratcTaskToken :: Lens' RespondActivityTaskCompleted Text

The taskToken of the ActivityTask.

taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

Response

data RespondActivityTaskCompletedResponse

Response constructor

respondActivityTaskCompletedResponse :: RespondActivityTaskCompletedResponse

RespondActivityTaskCompletedResponse constructor.