Network.AWS.SWF.RespondActivityTaskFailed

Description

Used by workers to tell the service that the ActivityTask identified by the taskToken has failed with reason (if specified). The reason and details appear in the ActivityTaskFailed event added to the workflow history.

A task is considered open from the time that it is scheduled until it is closed. Therefore a task is reported as open while a worker is processing it. A task is closed after it has been specified in a call to RespondActivityTaskCompleted , RespondActivityTaskCanceled , RespondActivityTaskFailed, or the task has timed out .

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

respondActivityTaskFailed

Arguments

:: Text
ratfTaskToken

-> RespondActivityTaskFailed

Creates a value of RespondActivityTaskFailed with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ratfReason - Description of the error that may assist in diagnostics.

ratfDetails - Detailed information about the failure.

ratfTaskToken - The taskToken of the ActivityTask . Important: taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

data RespondActivityTaskFailed

See: respondActivityTaskFailed smart constructor.

Request Lenses

ratfReason :: Lens' RespondActivityTaskFailed (Maybe Text)

Description of the error that may assist in diagnostics.

ratfDetails :: Lens' RespondActivityTaskFailed (Maybe Text)

Detailed information about the failure.

ratfTaskToken :: Lens' RespondActivityTaskFailed Text

The taskToken of the ActivityTask . Important: taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

Destructuring the Response

respondActivityTaskFailedResponse :: RespondActivityTaskFailedResponse

Creates a value of RespondActivityTaskFailedResponse with the minimum fields required to make a request.

data RespondActivityTaskFailedResponse

See: respondActivityTaskFailedResponse smart constructor.