Network.AWS.SWF.RespondDecisionTaskCompleted

Description

Used by deciders to tell the service that the DecisionTask identified by the taskToken has successfully completed. The decisions argument specifies the list of decisions made while processing the task.

A DecisionTaskCompleted event is added to the workflow history. The executionContext specified is attached to the event in the workflow execution history.

Access Control

If an IAM policy grants permission to use RespondDecisionTaskCompleted , it can express permissions for the list of decisions in the decisions parameter. Each of the decisions has one or more parameters, much like a regular API call. To allow for policies to be as readable as possible, you can express permissions on decisions as if they were actual API calls, including applying conditions to some parameters. For more information, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

respondDecisionTaskCompleted

Arguments

:: Text
rdtcTaskToken

-> RespondDecisionTaskCompleted

Creates a value of RespondDecisionTaskCompleted with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdtcDecisions - The list of decisions (possibly empty) made by the decider while processing this decision task. See the docs for the Decision structure for details.

rdtcExecutionContext - User defined context to add to workflow execution.

rdtcTaskToken - The taskToken from the DecisionTask . Important: taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

data RespondDecisionTaskCompleted

Input data for a TaskCompleted response to a decision task.

See: respondDecisionTaskCompleted smart constructor.

Request Lenses

rdtcDecisions :: Lens' RespondDecisionTaskCompleted [Decision]

The list of decisions (possibly empty) made by the decider while processing this decision task. See the docs for the Decision structure for details.

rdtcExecutionContext :: Lens' RespondDecisionTaskCompleted (Maybe Text)

User defined context to add to workflow execution.

rdtcTaskToken :: Lens' RespondDecisionTaskCompleted Text

The taskToken from the DecisionTask . Important: taskToken is generated by the service and should be treated as an opaque value. If the task is passed to another process, its taskToken must also be passed. This enables it to provide its progress and respond with results.

Destructuring the Response

respondDecisionTaskCompletedResponse :: RespondDecisionTaskCompletedResponse

Creates a value of RespondDecisionTaskCompletedResponse with the minimum fields required to make a request.

data RespondDecisionTaskCompletedResponse

See: respondDecisionTaskCompletedResponse smart constructor.