Network.AWS.SWF.StartWorkflowExecution

Description

Starts an execution of the workflow type in the specified domain using the provided workflowId and input data.

This action returns the newly started workflow execution.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagList.member.0 : The key is swf:tagList.member.0 .

tagList.member.1 : The key is swf:tagList.member.1 .

tagList.member.2 : The key is swf:tagList.member.2 .

tagList.member.3 : The key is swf:tagList.member.3 .

tagList.member.4 : The key is swf:tagList.member.4 .

taskList : String constraint. The key is swf:taskList.name .

workflowType.name : String constraint. The key is swf:workflowType.name .

workflowType.version : String constraint. The key is swf:workflowType.version .

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

startWorkflowExecution

Arguments

:: Text
sDomain

-> Text
sWorkflowId

-> WorkflowType
sWorkflowType

-> StartWorkflowExecution

Creates a value of StartWorkflowExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sTagList - The list of tags to associate with the workflow execution. You can specify a maximum of 5 tags. You can list workflow executions with a specific tag by calling ListOpenWorkflowExecutions or ListClosedWorkflowExecutions and specifying a TagFilter .

sTaskStartToCloseTimeout - Specifies the maximum duration of decision tasks for this workflow execution. This parameter overrides the defaultTaskStartToCloseTimout specified when registering the workflow type using RegisterWorkflowType . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

sLambdaRole - The IAM role to attach to this workflow execution.

sInput - The input for the workflow execution. This is a free form string which should be meaningful to the workflow you are starting. This input is made available to the new workflow execution in the WorkflowExecutionStarted history event.

sExecutionStartToCloseTimeout - The total duration for this workflow execution. This overrides the defaultExecutionStartToCloseTimeout specified when registering the workflow type. The duration is specified in seconds; an integer greater than or equal to 0 . Exceeding this limit causes the workflow execution to time out. Unlike some of the other timeout parameters in Amazon SWF, you cannot specify a value of NONE for this timeout; there is a one-year max limit on the time that a workflow execution can run.

sTaskList - The task list to use for the decision tasks generated for this workflow execution. This overrides the defaultTaskList specified when registering the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

sTaskPriority - The task priority to use for this workflow execution. This overrides any default priority that was assigned when the workflow type was registered. If not set, then the default task priority for the workflow type is used. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority. For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide .

sChildPolicy - If set, specifies the policy to use for the child workflow executions of this workflow execution if it is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This policy overrides the default child policy specified when registering the workflow type using RegisterWorkflowType . The supported child policies are: * TERMINATE – The child executions are terminated. * REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. * ABANDON – No action is taken. The child executions continue to run.

sDomain - The name of the domain in which the workflow execution is created.

sWorkflowId - The user defined identifier associated with the workflow execution. You can use this to associate a custom identifier with the workflow execution. You may specify the same identifier if a workflow execution is logically a restart of a previous execution. You cannot have two open workflow executions with the same workflowId at the same time. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

sWorkflowType - The type of the workflow to start.

data StartWorkflowExecution

See: startWorkflowExecution smart constructor.

Request Lenses

sTagList :: Lens' StartWorkflowExecution [Text]

The list of tags to associate with the workflow execution. You can specify a maximum of 5 tags. You can list workflow executions with a specific tag by calling ListOpenWorkflowExecutions or ListClosedWorkflowExecutions and specifying a TagFilter .

sTaskStartToCloseTimeout :: Lens' StartWorkflowExecution (Maybe Text)

Specifies the maximum duration of decision tasks for this workflow execution. This parameter overrides the defaultTaskStartToCloseTimout specified when registering the workflow type using RegisterWorkflowType . The duration is specified in seconds, an integer greater than or equal to 0 . You can use NONE to specify unlimited duration.

sLambdaRole :: Lens' StartWorkflowExecution (Maybe Text)

The IAM role to attach to this workflow execution.

sInput :: Lens' StartWorkflowExecution (Maybe Text)

The input for the workflow execution. This is a free form string which should be meaningful to the workflow you are starting. This input is made available to the new workflow execution in the WorkflowExecutionStarted history event.

sExecutionStartToCloseTimeout :: Lens' StartWorkflowExecution (Maybe Text)

The total duration for this workflow execution. This overrides the defaultExecutionStartToCloseTimeout specified when registering the workflow type. The duration is specified in seconds; an integer greater than or equal to 0 . Exceeding this limit causes the workflow execution to time out. Unlike some of the other timeout parameters in Amazon SWF, you cannot specify a value of NONE for this timeout; there is a one-year max limit on the time that a workflow execution can run.

sTaskList :: Lens' StartWorkflowExecution (Maybe TaskList)

The task list to use for the decision tasks generated for this workflow execution. This overrides the defaultTaskList specified when registering the workflow type. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

sTaskPriority :: Lens' StartWorkflowExecution (Maybe Text)

The task priority to use for this workflow execution. This overrides any default priority that was assigned when the workflow type was registered. If not set, then the default task priority for the workflow type is used. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority. For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide .

sChildPolicy :: Lens' StartWorkflowExecution (Maybe ChildPolicy)

If set, specifies the policy to use for the child workflow executions of this workflow execution if it is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This policy overrides the default child policy specified when registering the workflow type using RegisterWorkflowType . The supported child policies are: * TERMINATE – The child executions are terminated. * REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. * ABANDON – No action is taken. The child executions continue to run.

sDomain :: Lens' StartWorkflowExecution Text

The name of the domain in which the workflow execution is created.

sWorkflowId :: Lens' StartWorkflowExecution Text

The user defined identifier associated with the workflow execution. You can use this to associate a custom identifier with the workflow execution. You may specify the same identifier if a workflow execution is logically a restart of a previous execution. You cannot have two open workflow executions with the same workflowId at the same time. The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (u0000-u001f | u007f-u009f ). Also, it must not contain the literal string arn .

sWorkflowType :: Lens' StartWorkflowExecution WorkflowType

The type of the workflow to start.

Destructuring the Response

startWorkflowExecutionResponse

Arguments

:: Int
swersResponseStatus

-> StartWorkflowExecutionResponse

Creates a value of StartWorkflowExecutionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

swersRunId - The runId of a workflow execution. This ID is generated by the service and can be used to uniquely identify the workflow execution within a domain.

swersResponseStatus - -- | The response status code.

data StartWorkflowExecutionResponse

Specifies the runId of a workflow execution.

See: startWorkflowExecutionResponse smart constructor.

Response Lenses

swersRunId :: Lens' StartWorkflowExecutionResponse (Maybe Text)

The runId of a workflow execution. This ID is generated by the service and can be used to uniquely identify the workflow execution within a domain.

swersResponseStatus :: Lens' StartWorkflowExecutionResponse Int

- | The response status code.