Network.AWS.SWF.TerminateWorkflowExecution

Description

Records a WorkflowExecutionTerminated event and forces closure of the workflow execution identified by the given domain, runId, and workflowId. The child policy, registered with the workflow type or specified when starting this execution, is applied to any open child workflow executions of this workflow execution.

Important: If the identified workflow execution was in progress, it is terminated immediately.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED . For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide .

Creating a Request

terminateWorkflowExecution

Arguments

:: Text
tweDomain

-> Text
tweWorkflowId

-> TerminateWorkflowExecution

Creates a value of TerminateWorkflowExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tweReason - A descriptive reason for terminating the workflow execution.

tweRunId - The runId of the workflow execution to terminate.

tweChildPolicy - If set, specifies the policy to use for the child workflow executions of the workflow execution being terminated. This policy overrides the child policy specified for the workflow execution at registration time or when starting the execution. The supported child policies are: * TERMINATE – The child executions are terminated. * REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. * ABANDON – No action is taken. The child executions continue to run.

tweDetails - Details for terminating the workflow execution.

tweDomain - The domain of the workflow execution to terminate.

tweWorkflowId - The workflowId of the workflow execution to terminate.

data TerminateWorkflowExecution

See: terminateWorkflowExecution smart constructor.

Request Lenses

tweReason :: Lens' TerminateWorkflowExecution (Maybe Text)

A descriptive reason for terminating the workflow execution.

tweRunId :: Lens' TerminateWorkflowExecution (Maybe Text)

The runId of the workflow execution to terminate.

tweChildPolicy :: Lens' TerminateWorkflowExecution (Maybe ChildPolicy)

If set, specifies the policy to use for the child workflow executions of the workflow execution being terminated. This policy overrides the child policy specified for the workflow execution at registration time or when starting the execution. The supported child policies are: * TERMINATE – The child executions are terminated. * REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event. * ABANDON – No action is taken. The child executions continue to run.

tweDetails :: Lens' TerminateWorkflowExecution (Maybe Text)

Details for terminating the workflow execution.

tweDomain :: Lens' TerminateWorkflowExecution Text

The domain of the workflow execution to terminate.

tweWorkflowId :: Lens' TerminateWorkflowExecution Text

The workflowId of the workflow execution to terminate.

Destructuring the Response

terminateWorkflowExecutionResponse :: TerminateWorkflowExecutionResponse

Creates a value of TerminateWorkflowExecutionResponse with the minimum fields required to make a request.

data TerminateWorkflowExecutionResponse

See: terminateWorkflowExecutionResponse smart constructor.